% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_pick+plot.r
\name{adLinesToVector}
\alias{adLinesToVector}
\title{Ad Lines to a Vector Plot}
\description{
Ad various lines, i.e. graphical elements, to the plot of a 
vector like loading or regression vector.
}
\details{
XXX
}

