% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniset_functions.R
\name{ap2_settings_setup}
\alias{ap2_settings_setup}
\title{Perform Setup}
\usage{
ap2_settings_setup(path = NULL)
}
\arguments{
\item{path}{Character length one. The path to the folder where the settings-
home folder should be located. Defaults to \code{NULL}.}
}
\value{
Called for its side-effects, i.e. performing the required setup 
regarding the settings.R file.
}
\description{
Is performing the required setup regarding the settings.R file. 
Only has to be called once, but can be called repeatedly to conveniently 
change the location of the settings-home folder.
}
\details{
If argument \code{path} is left at its default \code{NULL}, the path 
to the desired folder can be selected interactively.
}
\seealso{
Other Helper Functions: 
\code{\link{ap2dme}()},
\code{\link{export_ap2_ToXlsx}()},
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{getstn}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()},
\code{\link{updateSettings}()}
}
\concept{Helper Functions}
