% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{ap2dme}
\alias{ap2dme}
\title{Download and Make Example Experiment}
\usage{
ap2dme(where, expName, ffs = FALSE, fdo = FALSE)
}
\arguments{
\item{where}{Character length one. The path where the experiment home folder
should be looked for. The experiment home folder, if already existing,  
has to end in \code{@home} in order to be recognized correctly.}

\item{expName}{Character length one, the name of the experiment.}

\item{ffs}{Logical. If the folder structure should be generated in all cases.
If set to \code{TRUE} and \code{where} is pointing to an already existing 
experiment home folder, the latter one will be deleted and a new folder 
structure will be created.}

\item{fdo}{Logical. If the download should always be enforced}
}
\value{
Used for its side effect, i.e. do (possibly) download example data 
and to create a working experiment folder structure. 
Returns \code{NULL} if no data were downloaded, \code{FALSE} if data 
should have been downloaded but it failed, and \code{TRUE} if data were 
possibly downloaded, the experiment folder 
structure was created successfully, and the data files were copied.
}
\description{
Checks if a dataset is present. If not, it is downloading a 
single example experiment from a server, generating the required folder 
structure and inserting the downloaded files into their appropriate folder.
}
\details{
The experiment home folder, if already 
present, has to end in \code{@home} in order to be recognized correctly.
If the experiment home folder does not exist it will be created.
Currently, this function is primary intended to be used in testing 
and to help create executable examples.
The downloaded folder containing all available example datasets is stored in 
\code{tempdir()}. If already present, it will not be downloaded again.
}
\seealso{
Other Helper Functions: 
\code{\link{ap2_settings_setup}()},
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{getstn}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()},
\code{\link{updateSettings}()}
}
\concept{Helper Functions}
