% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aquap2.r
\docType{package}
\name{aquap2}
\alias{aquap2}
\alias{aquap2-package}
\title{Aquaphotomics NIRS Analysis}
\description{
Methods and tools to analyse NIRS data from an Aquaphotomics perspective
}
\details{
The details.
}
\section{Maintainer}{
 Bernhard Pollner <bernhard.pollner@mac.com>
}

\section{Important Functions}{
 List functions here
}
\examples{
\dontrun{
# ad examples here
}
}
\author{
Bernhard Pollner, Zoltan Kovacs
}

