% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniset_functions.R
\name{aquap2_handover_to_uniset}
\alias{aquap2_handover_to_uniset}
\title{Handover to Uniset}
\usage{
aquap2_handover_to_uniset()
}
\description{
Is handing over values to package uniset. These values are
required for the correct functioning of the dynamic settings file system
provided by \code{\link[uniset]{uniset}}.
}
