% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_classes.r
\name{aquap_cpt}
\alias{aquap_cpt}
\title{Class 'aquap_cpt'}
\description{
Used to generate datasets.
}
\details{
The 'cpt' object is in the '@cpt' slot of the \code{\link{aquap_cube}}
}
\section{Slots}{

\describe{
\item{\code{len}}{Numeric length one, the number of rows of the comparison pattern.}

\item{\code{splitVars}}{A list with two elements:
\itemize{
 \item "classes" A data frame with the column names of the variables used for 
 splitting. 
 \item "values" A data frame with the values of the columns (variables) used 
 for splitting.
}}

\item{\code{wlSplit}}{A list with to numerics (wavelength 'from' and 'to') in each 
element.}

\item{\code{smoothing}}{A logical vector.}

\item{\code{noise}}{A logical vector.}
}}
\seealso{
\code{\link{gdmm}}

Other Class documentations: \code{\link{aquap_cube}},
  \code{\link{aquap_set}}
}

