% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_classes.r
\name{aquap_cube}
\alias{aquap_cube}
\title{Class 'aquap_cube'}
\description{
Holds all the statistical models / calculations that were 
performed on the split-variations of the dataset by the function 
\code{\link{gdmm}} in a list.
}
\details{
Each element of the list in aquap_cube is an object of class 
\code{\link{aquap_set}}.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list with one object of class \code{\link{aquap_set}} in each 
element.}

\item{\code{metadata}}{An object of class 'aquap_md' (what is list)}

\item{\code{anproc}}{An object of class 'aquap_ap' (what is a list)}

\item{\code{cp}}{A data frame with the 'comparison pattern', i.e. a description of the 
split-variations of the dataset in a well readable form. This slots gets printed 
to the screen when you just type the name of a cube-object. (method 'show')}

\item{\code{cpt}}{An object of class \code{\link{aquap_cpt}}, what is basically just an 
other version of the data in 'cp' for internal use.}
}}

\seealso{
\code{\link{gdmm}}

Other Class documentations: \code{\link{aquap_cpt}},
  \code{\link{aquap_set}}
}
