% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.r, R/doc_methods.r
\docType{methods}
\name{[,aquap_data-method}
\alias{-,aquap_data,aquap_data-method}
\alias{[,aquap_data-method}
\alias{aquap_data-methods}
\title{aquap_data methods}
\usage{
\S4method{[}{aquap_data}(x, i)

\S4method{-}{aquap_data,aquap_data}(e1, e2)
}
\arguments{
\item{x}{An object of class 'aquap_data'}

\item{i}{subsricpting indices for rows}

\item{e1}{Object of class 'aquap_data'}

\item{e2}{Object of class 'aquap_data'}
}
\description{
Available methods for objects of class 'aquap_data', used for
subtractions, subscripting etc. of datasets.
}
\details{
\describe{
\item{Subscripting via '[]'}{For subscripting via \code{'[]'}, only values 
for rownumbers are accepted.}
\item{Subtractions via '-'}{Use the subtraction via \code{'-'} to subtract the 
NIR-data in datasets that have the same structure, i.e. the same header, and 
the same wavelengths in the NIR. You can either subtract two equally long 
datasets, i.e. two datasets having the exactly same number of rows, or you can
subtract a dataset having exactly one row from a full dataset. The latter can
be useful for e.g. subtracting the averaged spectra of a single class from the
complete dataset - see examples.}
}
}
\note{
\code{drop} is always set to \code{FALSE} for substrictping via 
\code{'[]'}.
}
\examples{
\dontrun{
 dataset <- gfd()
 ds <- dataset[1:5,]
 ds <- dataset[1:5] # the same as above
	dSub <- dataset - do_sgolay(dataset, n=71) # apply smoothing and subtract
 plot(dSub, pg.where="", pg.main=" | smoothed subtracted")
}
}
\seealso{
Other Data pre-treatment functions: \code{\link{do_avg}},
  \code{\link{do_emsc}}, \code{\link{do_msc}},
  \code{\link{do_sgolay}}, \code{\link{dpt_modules}},
  \code{\link{selectWls}}, \code{\link{ssc}}
}

