% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doc_classes.r
\name{aquap_set}
\alias{aquap_set}
\title{Class 'aquap_set'}
\description{
Holds a dataset and all the statistical models / calculations that
were performed on this specific dataset.
}
\details{
The 'aquap_set' is in each list-element of the \code{\link{aquap_cube}}
object.
}
\section{Slots}{

\describe{
\item{\code{dataset}}{The specific dataset used to calculate the models}

\item{\code{idString}}{A character length one describing the split-parameter of the
dataset.}

\item{\code{pca}}{The PCA model.}

\item{\code{plsr}}{The PLSR model(s).}

\item{\code{simca}}{The SIMCA model(s).}

\item{\code{aquagr}}{All data related to the calculation of Aquagrams.}
}}
\seealso{
\code{\link{gdmm}}

Other Class documentations: \code{\link{aquap_cpt}};
  \code{\link{aquap_cube}}
}

