% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_aqg.r
\name{calc_aqg_modes}
\alias{calc_aqg_modes}
\title{Calculate Aquagram - Modes}
\description{
The following values can be provided to the argument \code{aqg.mod}
in \code{getap}, specifiying what type of Aquagram should be calculated.
}
\details{
XXX
}
\section{Possible Values}{
 Possible Aquagram modes are
\itemize{
\item \code{classic}: The classical aquagram - smoothed, MSC, normalized
\item \code{classic-diff}: same as above + one group subtracted from all the
others.
\item \code{sfc}: "Scaled, foreign center": smoothed, MSC, centered on the mean of
selected calibration data, scaled.
\item \code{sfc-diff}: same as above + one group subtracted from all the others.
\item \code{aucs}: "Area Under Curve Stabilized": the area under the curve (auc)
from the spectra in all the 12/15 coordinates in a set of calibration data
at varying temperature is calculated, then divided by the total auc. Then
the smalles and biggest value is taken as 0% and 100%. The same auc is
calculated for every sample, and its value in percent ... XXX
\item \code{aucs-diff}: same as above + one group subtracted from all the others.
\item \code{aucs.tn}: aucs + temperature normalization: .... XXX the auc from a
sample at Texp gets subtracted from all the auc of the samples.
\item \code{aucs.tn-diff}: same as above + one group subtracted from all the
others
\item \code{aucs.tn.dce}: same as aucs.tn, but the scale calculated to show
degrees celsius equivalent
\item \code{aucs.tn.dce-diff}: same as above + one group subtracted from all the
others
\item \code{aucs.dce}: same as aucs, but the scale calculated to show degrees
celsius equivalent
\item \code{aucs.dce-diff}: same as above + one group subtracted from all the
others
}
}
\references{
XXX the Aquagram paper XXX
}
\seealso{
\code{\link{calc_aqg_args}}
}

