% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_classification.r
\name{calc_discrimAnalysis_args}
\alias{calc_discrimAnalysis_args}
\title{Calculate Discriminant Analysis - Arguments}
\arguments{
\item{do.da}{Logical. If used in \code{getap}, if classification via 
discriminant analysis (\code{\link[MASS]{lda}}, \code{\link[MASS]{qda}}, 
\code{\link[mda]{fda}}, \code{\link[mclust]{MclustDA}}) should be performed 
in the given dataset.}

\item{da.type}{Character vector. The type of discriminant analysis (DA) to 
perform; possible values (one or more) are: 
\code{'lda', 'qda', 'fda', 'mclustda'}:
\itemize{
   \item \code{lda} Linear DA using 
   \code{\link[MASS]{lda}}.
   \item \code{qda} Quadratic DA using 
   \code{\link[MASS]{qda}}.
   \item \code{fda} Flexible DA using 
   \code{\link[mda]{fda}}.
   \item \code{mclustda} DA based on Gaussian finite 
   mixture modeling using \code{\link[mclust]{MclustDA}}.
 }}

\item{da.classOn}{Character vector. One or more class variables to define the 
grouping used for classification.}

\item{da.testCV}{Logical, if the errors of the test-data should be crossvalidated. 
If set to true, CV and testing is repeated in alternating datasets. See below.}

\item{da.percTest}{Numeric length one. The percentage of the dataset that should 
be set aside for testing the models; these data are never seen during training 
and crossvalidation.}

\item{da.cvBootCutoff}{If the number of observations within the smallest 
subgroup defined by the classification grouping variable is EQUAL or HIGHER then 
\code{.cvBootCutoff}, the crossvalidation is done via splitting the training 
data in the number of segments defined in \code{.valid} (see below), otherwise 
the crossvalidation is done via bootstrap resampling, with the number of 
bootstrap iterations resulting from the multiplication by the number of 
observations in this smallest subgroup with \code{.cvBootFactor}. Set the latter 
to 1 to NEVER perform the CV of the training data via bootstrap, or, for a 
global solution, set the parameter \code{cl_gen_neverBootstrapForCV} in the 
settings file to \code{TRUE}.}

\item{da.cvBootFactor}{The factor used to multiply the number of observations 
within the smallest subgroup defined by the classification grouping variable 
with, resulting in the number of iterations of a possible bootstrap 
crossvalidation of the trainign data -- see \code{.cvBootCutoff}.}

\item{da.valid}{The number of segments the training data should be divided 
into in case of a "traditional" crossvalidation of the training data; see above.}
}
\description{
The following parameters can be used in the \code{...} argument in 
function \code{\link{getap}}, also within function \code{\link{gdmm}}, to 
override the values in the analysis procedure file and so to modify the 
calculation of discriminant analysis models - see examples.
\describe{
\item{\code{getap(...)}}{ }
\item{\code{gdmm(dataset, ap=getap(...))}}{ }
}
}
\details{
For a list of all parameters that can be used in the \code{...} 
argument in \code{\link{getap}} and in the \code{\link{plot}} functions 
please see \code{\link{anproc_file}}.
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{gdmm}}, \code{\link{siWlg}} for reducing the number of 
wavelengths in a dataset

Other Calc. arguments: \code{\link{calc_NNET_args}},
  \code{\link{calc_SVM_args}}, \code{\link{calc_aqg_args}},
  \code{\link{calc_pca_args}}, \code{\link{calc_pls_args}},
  \code{\link{calc_randomForest_args}},
  \code{\link{calc_sim_args}}, \code{\link{split_dataset}}

Other Classification functions: \code{\link{calc_NNET_args}},
  \code{\link{calc_SVM_args}},
  \code{\link{calc_randomForest_args}}

Other DA documentation: \code{\link{plot_DA}},
  \code{\link{plot_discrimAnalysis_args}}
}
