% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_classification.r
\name{calc_NNET_args}
\alias{calc_NNET_args}
\title{Calculate NNET - Arguments}
\arguments{
\item{do.nnet}{Logical. If used in \code{getap}, if classification via 
artificial neural networks (\code{\link[nnet]{nnet}}) should be performed 
in the  given dataset.}

\item{nnet.classOn}{Character vector. One or more class variables to define the 
grouping used for classification.}

\item{nnet.testCV}{Logical, if the errors of the test-data should be crossvalidated. 
If set to true, CV and testing is repeated in alternating datasets. See below.}

\item{nnet.percTest}{Numeric length one. The percentage of the dataset that should 
be set aside for testing the models; these data are never seen during training 
and crossvalidation.}

\item{nnet.cvBootCutoff}{If the number of observations within the smallest 
subgroup defined by the classification grouping variable is EQUAL or HIGHER then 
\code{.cvBootCutoff}, the crossvalidation is done via splitting the training 
data in the number of segments defined in \code{.valid} (see below), otherwise 
the crossvalidation is done via bootstrap resampling, with the number of 
bootstrap iterations resulting from the multiplication by the number of 
observations in this smallest subgroup with \code{.cvBootFactor}. Set the latter 
to 1 to NEVER perform the CV of the training data via bootstrap, or, for a 
global solution, set the parameter \code{cl_gen_neverBootstrapForCV} in the 
settings file to \code{TRUE}.}

\item{nnet.cvBootFactor}{The factor used to multiply the number of observations 
within the smallest subgroup defined by the classification grouping variable 
with, resulting in the number of iterations of a possible bootstrap 
crossvalidation of the trainign data -- see \code{.cvBootCutoff}.}

\item{nnet.valid}{The number of segments the training data should be divided 
into in case of a "traditional" crossvalidation of the training data; see above.}

\item{nnet.pcaRed}{Logical, if variable reduction via PCA should be applied; if 
TRUE, the subsequent classifications are performed on the PCA scores, see
\code{nnet.pcaNComp} below.}

\item{nnet.pcaNComp}{Character or integer vector. Provide the character "max" 
to use the maximum number of components (i.e. the number of observations minus 
1), or an integer vector specifying the components resp. their scores to be 
used for nnet classification.}
}
\description{
The following parameters can be used in the \code{...} argument in 
function \code{\link{getap}}, also within function \code{\link{gdmm}}, to 
override the values in the analysis procedure file and so to modify the 
calculation of articial neural network (NNET/ANN) models - see examples.
\describe{
\item{\code{getap(...)}}{ }
\item{\code{gdmm(dataset, ap=getap(...))}}{ }
}
}
\details{
For a list of all parameters that can be used in the \code{...} 
argument in \code{\link{getap}} and in the \code{\link{plot}} functions 
please see \code{\link{anproc_file}}.
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{gdmm}}, \code{\link{siWlg}} for reducing the number of 
wavelengths in a dataset

Other Calc. arguments: \code{\link{calc_SVM_args}},
  \code{\link{calc_aqg_args}},
  \code{\link{calc_discrimAnalysis_args}},
  \code{\link{calc_pca_args}}, \code{\link{calc_pls_args}},
  \code{\link{calc_randomForest_args}},
  \code{\link{calc_sim_args}}, \code{\link{split_dataset}}

Other Classification functions: \code{\link{calc_SVM_args}},
  \code{\link{calc_discrimAnalysis_args}},
  \code{\link{calc_randomForest_args}}

Other NNET documentation: \code{\link{plot_NNET_args}},
  \code{\link{plot_NNET}}
}
