% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.r
\name{calc_pca_args}
\alias{calc_pca_args}
\title{Calculate PCA - Arguments}
\description{
The following parameters can be used in the \code{...} argument in 
function \code{\link{getap}}, also within function \code{\link{gdmm}}, to 
override the values in the analysis procedure file and so to modify the 
calculation of PCA - see examples.

\code{getap(...)}
 
\code{gdmm(dataset, ap=getap(...))}

@param do.pca Logical. If used in \code{getap}, if a PCA should be calculated 
 with a given dataset.
}
\details{
For a list of all parameters that can be used in the \code{...} 
argument in \code{\link{getap}} and in the \code{\link{plot}} functions 
please see \code{\link{anproc_file}}.
}
\section{Note}{
 Calculation of PCA is done with the function 
\code{\link[ChemometricsWithR]{PCA}}
}
\examples{
\dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset, getap(do.pca=TRUE))
}
}
\seealso{
\code{\link{gdmm}}

Other Calc. arguments: \code{\link{calc_aqg_args}},
  \code{\link{calc_pls_args}}, \code{\link{calc_sim_args}},
  \code{\link{split_dataset}}

Other PCA documentation: \code{\link{plot_pca_args}}
}

