% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simca.r
\name{calc_sim_args}
\alias{calc_sim_args}
\title{Calculate SIMCA - Arguments}
\arguments{
\item{do.sim}{Logical. If used in \code{getap}, if SIMCA models of the dataset 
should be calculated.}

\item{sim.vars}{NULL or character vector. Which variables should be 
used to group the data. Set to NULL for using all available class-variables, 
or provide a character vector with the column names of class variables to group
the data along those for calculating SIMCA models.}

\item{sim.K}{Numeric length one. The number of components used for calculating 
the SIMCA models. In mode 'robust' leave at '0' for automatic detection of 
optimal number of components. [It is a capital 'K' in the argument.]}
}
\description{
The following parameters can be used in the \code{...} argument in 
function \code{\link{getap}}, also within function \code{\link{gdmm}}, to 
override the values in the analysis procedure file and so to modify the 
calculation of SIMCA models - see examples.

\code{getap(...)}
 
\code{gdmm(dataset, ap=getap(...))}
}
\details{
For a list of all parameters that can be used in the \code{...} 
argument in \code{\link{getap}} and in the \code{\link{plot}} functions 
please see \code{\link{anproc_file}}.
}
\section{Note}{
 Calculation of SIMCA models is done with the function 
\code{\link[rrcovHD]{RSimca}} /  \code{\link[rrcovHD]{CSimca}} and with code 
adapted from the manual-page of the Chemometrics-software "Pirouette" XXX.
}
\examples{
\dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset, getap(do.sim=TRUE))
 cube <- gdmm(dataset, getap(do.sim=TRUE, sim.K=4))
}
}
\seealso{
\code{\link{gdmm}}

Other Calc. arguments: \code{\link{calc_aqg_args}},
  \code{\link{calc_pca_args}}, \code{\link{calc_pls_args}},
  \code{\link{split_dataset}}

Other SIMCA documentation: \code{\link{plot_sim_args}},
  \code{\link{plot_sim}}
}

