% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_classification.r
\name{classification_helpers}
\alias{classification_helpers}
\title{Classification - General Considerations}
\description{
General remarks regarding the use of the classification functions 
in package aquap2.
\itemize{
\item Except the svm classification, most methods do not work well in fat data 
matrices, so if there are more variables (wavelengths) than observations.
For an easy reduction of the number of wavelengths we provide, especially in 
the Aquaphotomics context, a special function to reduce the number of 
wavelengths in a dataset (as produced by \code{\link{gfd}}): function 
\code{\link{siWlg}} can isolate any custom group of wavelengths from the dataset, 
or, if left at its default, is isolating the ranges of the 12 water matrix 
coordinates within the 1st overtone. For further data reduction, this groups 
can be averaged within as well.
}
}
\seealso{
Other Classification Helpers: \code{\link{siWlg}}
}
