% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_modifyDatasets.r
\name{combineVariable}
\alias{combineVariable}
\title{Combine Variable}
\usage{
combineVariable(dataset, sourceVars = NULL, name = NULL, sep = "_")
}
\arguments{
\item{dataset}{The standard dataset as produced by \code{\link{gfd}}.}

\item{sourceVars}{Character length two or more. The name of the variables where 
values should be copied from.}

\item{name}{Character length one. The name of the new variable.}

\item{sep}{Character length one. A character to separate the individual 
values in the new variable. Set to '""' (empty string) to have no separation 
between values.}
}
\value{
An object of class 'aquap_data' with a new variable in the header.
}
\description{
Combine values from several variables into a single value in 
a mew variable.
}
\details{
The resulting new variable can only be a class-variable.
}
\examples{
\dontrun{
fd <- gfd() # load a dataset
fdNew <- combineVariable(fd, c("C_Foo", "C_Bar"), "FooBar")
fdNew <- combineVariable(fd, c("C_Foo", "C_Bar"), "FooBar", sep="*")
	# is combining the values of the two variables 'C_Foo' and 'C_Bar' into a new 
 # variable called 'C_FooBar'.
}
}
\seealso{
calculateVariable

Other dataset modification functions: 
\code{\link{[,aquap_data-method}},
\code{\link{calculateVariable}()},
\code{\link{generateMergeLabels,aquap_data,aquap_data,character,character-method}},
\code{\link{mergeDatasets,aquap_data,aquap_data,missing-method}}
}
\concept{dataset modification functions}
