% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_noise.r
\name{do_addNoise}
\alias{do_addNoise}
\title{Add Noise to Dataset}
\usage{
do_addNoise(dataset, noiseFile = "def", md = getmd())
}
\arguments{
\item{dataset}{An object of class 'aquap_data' as generated by \code{\link{gfd}}}

\item{noiseFile}{Character length one. The name of the corresponding noise data 
file residing in the AQUAP2SH folder. Leave at the default 'def' to read in the 
value from the settings.r file (parameter \code{noi_noiseDataFilename}), or 
provide the name of an existing R-data file in your specified AQUAP2SH folder 
to use this as source for the noise data. If a value other than "def" is 
provided, this is overriding the value of "noiseFileName" in the metadata file.
Please see \code{\link{noise_procedures}} for more information.}

\item{md}{The metadata from where the name of the noise-file is taken.}
}
\value{
The dataset with added noise.
}
\description{
Manually add noise to a provided dataset.
}
\details{
Select the noise-mode in the settings.r file (parameter 
\code{noi_addMode}). You need to have an R-data file with noise-spectra in 
your settings-home folder, please refer to \code{\link{noise_procedures}} for 
further information.
}
\examples{
\dontrun{
fd <- gfd()
fdNoise <- do_addNoise(fd)
}
}
\seealso{
\code{\link{noise_procedures}}

Other Noise procedures: \code{\link{genNoiseRecExp}},
  \code{\link{noise_procedures}}

Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_avg}}, \code{\link{do_blowup}},
  \code{\link{do_detrend}}, \code{\link{do_emsc}},
  \code{\link{do_gapDer}}, \code{\link{do_msc}},
  \code{\link{do_resampleNIR}}, \code{\link{do_sgolay}},
  \code{\link{dpt_modules}}, \code{\link{selectWls}},
  \code{\link{ssc}}
}
