% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_avg}
\alias{do_avg}
\title{Average spectra in a dataset}
\usage{
do_avg(dataset)
}
\arguments{
\item{dataset}{An object of class 'aquap_data" as produced by 
\code{\link{gfd}} or as can be extracted from a 'cube' via 
\code{\link{getcd}}.}
}
\value{
A standard dataset with a single row.
}
\description{
Calculate the average spectrum of all the given spectra in a 
dataset.
}
\details{
The average spectrum is returned in a single row with the 
header-data taken from the first row of the original dataset.
}
\examples{
\dontrun{
fd <- gfd() # get the full dataset
fd <- selectWls(fd, 1100, 1800)
fd_avg <- do_avg(fd)
fd_avg2 <- do_avg(ssc(fd, C_water=="MQ")) # assumes the existence of column "C_water"
plot(fd_avg, pg.where="", pg.main="| averaged")
plot(fd - fd_avg, pg.where="", pg.main="| avg subtracted")
}
}
\seealso{
\code{\link{getcd}}

Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_addNoise}}, \code{\link{do_detrend}},
  \code{\link{do_emsc}}, \code{\link{do_gapDer}},
  \code{\link{do_msc}}, \code{\link{do_resampleNIR}},
  \code{\link{do_sgolay}}, \code{\link{dpt_modules}},
  \code{\link{selectWls}}, \code{\link{ssc}}
}

