% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_blowup}
\alias{do_blowup}
\title{Increase the Numbers of Observation in Dataset}
\usage{
do_blowup(dataset, tn = "x10", an = "20\%", grp = NULL, cst = TRUE)
}
\arguments{
\item{dataset}{An object of class 'aquap_data' as produced e.g. by 
\code{\link{gfd}}.}

\item{tn}{Numeric or character length one. If numeric, \code{tn} is denoting 
the target-number, i.e. the desired number of observations in the expanded 
dataset. If a grouping is provided via \code{grp}, the target number is 
defining the desired number of observations within each subgroup obtained 
by the grouping. If \code{tn} is a character it hast to be in the format 
\code{xN}, with \code{N} being a positive integer. In this case it means 
an N-fold increase of the dataset resp. of each subgroup as defined by the 
grouping in \code{grp}. Defaults to \code{x10}.}

\item{an}{Numeric or character length one. If numeric, \code{an} is denoting 
the "average-number", i.e. the number of random observations resp. their spectra 
that are averaged together into a new spectrum. If \code{an} is a character it 
has to be in the format \code{N\%}, with \code{N} being a positive integer. In 
this case it denotes the percentage of observations in the dataset resp. in 
each of the subgroups possibly obtained via \code{grp} that should be averaged 
together into a new observation. Defaults to \code{20\%}.}

\item{grp}{Character. One ore more valid class-variable names that should be 
used to from subgroups within the dataset. Similar to the \code{spl.var} 
argument in the analysis procedure.}

\item{cst}{Logical. If consecutive scans should always be kept together. 
NOT YET IMPLEMENTED - value is ignored, and consec. scans are NOT kept together.}
}
\value{
The dataset with increased numbers of observations.
}
\description{
Use random observations of the provided dataset (within a possible 
grouping), calculate their average and add the resulting spectrum as new 
observation to the dataset.
}
\details{
The provenience of each observation is marked in an additional column 
in the header named `blowup`, where the character `orig` denotes an original 
observation, while the character `artif` is marking an artificially generated 
observation.
}
\examples{
\dontrun{
fd <- gfd()
fdPlus <- do_blowup(fd, tn="x4", an=4)
fdPlus <- do_blowup(fd, tn="x8", an="10\%", grp=c("C_Foo", "C_Bar")) 
fdPlus <- do_blowup(fd, tn=1000, an=5)
}
}
\seealso{
Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_addNoise}}, \code{\link{do_avg}},
  \code{\link{do_detrend}}, \code{\link{do_emsc}},
  \code{\link{do_gapDer}}, \code{\link{do_msc}},
  \code{\link{do_resampleNIR}}, \code{\link{do_sgolay}},
  \code{\link{dpt_modules}}, \code{\link{selectWls}},
  \code{\link{ssc}}
}
