% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_detrend}
\alias{do_detrend}
\title{Perform De-Trend}
\usage{
do_detrend(dataset, src = NULL, trg = "src")
}
\arguments{
\item{dataset}{An object of class 'aquap_data' as produced e.g. by 
\code{\link{gfd}}.}

\item{src}{'source'; the wavelength-range from where the values for de-trend 
should be calculated (i.e. the linear models). Leave at the default NULL 
to use the full range of wavelengths in the provided dataset, or provide 
a numeric length two to use this wavelength range for calculating the values 
for de-trend.}

\item{trg}{'target'; character length one or numeric length two. The wavelengths
where the de-trend should be applied. If left at the default 'src' the same 
wavelength as specified in \code{src} is used. Possible values are: 
\describe{
\item{src}{Apply the de-trend to the same wavelength-range as specified in 
argument \code{src} (the default).}
\item{all}{Apply the de-trend to all the wavelengths in the provided dataset.}
\item{Numeric length two}{Provide a numeric length two to the argument 
\code{trg} to apply the de-trend only to this wavelength range.}
}}
}
\value{
The transformed dataset
}
\description{
Perform de-trending of the given dataset. For each observation, 
the linear model \code{lm(absorbance ~ wavelength)} is calculated. The 
coefficients of the resulting model are then used to modify the absorbance 
values after the formula 
\code{newAbsorbance = absorbance - intercept - k*wavelength}. It is possible 
to separately specify the source and the target of the  de-trend operation.
}
\details{
Via the arguments \code{src} ('source') and \code{trg} ('target') 
it is possible to specify separately from which wavelength-range the values 
for the de-trend should be taken resp. to which wavelength-range the 
resulting de-trend should be applied. Please see documentation for the 
arguments \code{src} and \code{trg} for further details. If the target 
wavelengths are only a part of the whole wavelength-range in the dataset, 
the de-trend will be applied only to this wavelengths, and the rest of the 
wavelengths will remain unchanged. Abrupt steps in the resulting spectra can 
be the result of this. If both arguments \code{src} and \code{trg} are left 
at their defaults, the full range of wavelengths in the dataset is used for 
calculating the de-trend values, i.e. the linear models, and the resulting 
de-trend is also applied to the full range of wavelengths present in the 
dataset.
}
\examples{
\dontrun{
fd <- gfd()
fdDT <- do_detrend(fd) # use the whole wavelength range of 'fd' as source and
# target for the de-trend
plot(fd)
plot(fdDT)
###
fdc <- selectWls(fd, 1300, 1600)
plot(fdc)
plot(do_detrend(fdc)) # whole range as source and target
plot(do_detrend(fdc, src=c(1400, 1500))) # same target as soruce
plot(do_detrend(fdc, src=c(1400, 1500), trg="all")) # apply to full range
plot(do_detrend(fdc, src=c(1400, 1500), trg=c(1300, 1600))) # same as above 
plot(do_detrend(fdc, src=c(1300, 1400), trg=c(1380, 1580))) 
}
}
\seealso{
Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_avg}}, \code{\link{do_emsc}},
  \code{\link{do_gapDer}}, \code{\link{do_msc}},
  \code{\link{do_sgolay}}, \code{\link{dpt_modules}},
  \code{\link{selectWls}}, \code{\link{ssc}}

Other dpt modules documentation: \code{\link{do_emsc}},
  \code{\link{do_gapDer}}, \code{\link{do_msc}},
  \code{\link{do_sgolay}}, \code{\link{do_snv}},
  \code{\link{dpt_modules}}
}

