% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_emsc}
\alias{do_emsc}
\title{Perform EMSC}
\usage{
do_emsc(dataset, vecLoad = NULL, exportModel = FALSE)
}
\arguments{
\item{dataset}{An object of class 'aquap_data" as produced by 
\code{\link{gfd}} or as can be extracted from a 'cube' via 
\code{\link{getcd}}.}

\item{vecLoad}{A data frame x (\code{ncol(x) <= 2}) with one or two loading 
vectors or one regression vector.}

\item{exportModel}{Logical. If a possible model should be stored in the set,
(leave at \code{FALSE}).}
}
\value{
Returns the dataset with the transformed NIR data.
}
\description{
Performs EMSC with up to two signals on the provided dataset
}
\details{
A data frame with one or two loadings or with one regression vector
can be used as input to (try to) remove the effect on the data described by 
said loadings / regr. vector. For example, from a pca-model the first and 
second loading vector can be extracted with \code{pcaModel$loadings[, c(1,2)]}.
}
\examples{
\dontrun{
fd <- gfd()
cu <- gdmm(fd, getap(do.pca=TRUE)) # assumes no split
loadings <- getcm(cu, 1, what="pca")$loadings[, c(1,2)]
fd_emsc <- do_emsc(fd, loadings)
}
}
\seealso{
\code{\link{getcm}} for easy extraction of single models where 
loading vectors or a regression vector can be obtained.

Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_addNoise}}, \code{\link{do_avg}},
  \code{\link{do_blowup}}, \code{\link{do_detrend}},
  \code{\link{do_gapDer}}, \code{\link{do_msc}},
  \code{\link{do_resampleNIR}}, \code{\link{do_sgolay}},
  \code{\link{dpt_modules}}, \code{\link{selectWls}},
  \code{\link{ssc}}

Other dpt modules documentation: \code{\link{do_detrend}},
  \code{\link{do_gapDer}}, \code{\link{do_msc}},
  \code{\link{do_sgolay}}, \code{\link{do_snv}},
  \code{\link{dpt_modules}}
}
\concept{Data pre-treatment functions}
\concept{dpt modules documentation}
