% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_gapDer}
\alias{do_gapDer}
\title{Perform gap-segment derivatives}
\usage{
do_gapDer(dataset, m = 1, w = 1, s = 1, deltaW, exportModel = FALSE)
}
\arguments{
\item{dataset}{An object of class 'aquap_data" as produced by 
\code{\link{gfd}} or as can be extracted from a 'cube' via 
\code{\link{getcd}}.}

\item{m}{Numeric length one. The order of the derivative, can be between 1 and 
4. Default is 1.}

\item{w}{Numeric length one. The filter length (should be odd), ie. the spacing 
between points 
over which the derivative is computed. Default is 1.}

\item{s}{Numeric length one. The segment size, i.e. the range over which the 
points are averaged.Default is 1.}

\item{deltaW}{Numeric length one. The sampling interval / band spacing. Default 
is 1.}

\item{exportModel}{Logical. If a possible model should be stored in the set,
(leave at \code{FALSE}).}
}
\description{
Performs gap-segment derivatives on the provided dataset. The 
behaviour of the filter can be modified via its arguments, please see also 
the documentation for \code{\link[prospectr]{gapDer}}.
}
\details{
The first column of the wavelengths and the last one get doubled in 
order to have the same number of wavelengths in the resulting dataset as in 
the original dataset. The underlying function is \code{\link[prospectr]{gapDer}}.
}
\section{Note}{

The documentation for the parameters was mostly taken from 
\code{\link[prospectr]{gapDer}} by Antoine Stevens.
}

\examples{
\dontrun{
fd <- gfd()
fd_gsd <- do_gapDer(fd)
fd_gsd2 <- do_gapDer(fd, 1,11,13,1)
plot(fd_gsd - fd_gsd2)
}
}
\seealso{
\code{\link[prospectr]{gapDer}}

Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_addNoise}}, \code{\link{do_avg}},
  \code{\link{do_detrend}}, \code{\link{do_emsc}},
  \code{\link{do_msc}}, \code{\link{do_resampleNIR}},
  \code{\link{do_sgolay}}, \code{\link{dpt_modules}},
  \code{\link{selectWls}}, \code{\link{ssc}}

Other dpt modules documentation: \code{\link{do_detrend}},
  \code{\link{do_emsc}}, \code{\link{do_msc}},
  \code{\link{do_sgolay}}, \code{\link{do_snv}},
  \code{\link{dpt_modules}}
}
