% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_msc}
\alias{do_msc}
\title{Perform MSC}
\usage{
do_msc(dataset, reference = NULL)
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{reference}{XXX}
}
\value{
Returns the transformed dataset
}
\description{
Performs MSC on the provided data
}
\details{
The underlying function is \code{\link[pls]{msc}}.
}
\seealso{
Other Data pre-treatment functions: \code{\link{do_sgolay}},
  \code{\link{selectWls}}, \code{\link{ssc}}
}

