% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_resampleNIR}
\alias{do_resampleNIR}
\title{Resample data to new Wavelengths}
\usage{
do_resampleNIR(
  dataset,
  targetWls = NULL,
  tby = 0.5,
  method = getstn()$gen_resample_method
)
}
\arguments{
\item{dataset}{An object of class 'aquap_data' as produced e.g. by 
\code{\link{gfd}}.}

\item{targetWls}{Numeric vector, the target wavelengths. If left at the 
default \code{NULL}, an evenly spaced vector from the lowest to the 
highest available wavelength is automatically generated as the target 
wavelengths to be resampled to.}

\item{tby}{Numeric length one, the spacing in nm of the automatically 
generated target wavelength. Only applies if \code{targetWls} is left at 
\code{NULL}.}

\item{method}{The resampling method. For details see 
\code{\link[pracma]{interp1}}. Defaults to 'cubic'; the default can be 
changed in the settings.r file (parameter \code{gen_resample_method}).}
}
\value{
The resampled dataset.
}
\description{
Resample the data in the provided dataset to new, possibly 
evenly spaced, wavelengths.
}
\details{
If nothing is provided for the argument \code{targetWls}, an evenly 
spaced vector from the lowest to the highest available wavelength 
is automatically generated as the target wavelengths to be resampled to.
}
\examples{
\dontrun{
fd <- gfd()
fdR <- do_resampleNIR(fd)
}
}
\seealso{
Other Data pre-treatment functions: 
\code{\link{[,aquap_data-method}},
\code{\link{do_addNoise}()},
\code{\link{do_avg}()},
\code{\link{do_blowup}()},
\code{\link{do_detrend}()},
\code{\link{do_emsc}()},
\code{\link{do_gapDer}()},
\code{\link{do_msc}()},
\code{\link{do_sgolay}()},
\code{\link{dpt_modules}},
\code{\link{selectWls}()},
\code{\link{ssc}()}
}
\concept{Data pre-treatment functions}
