% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_sgolay}
\alias{do_sgolay}
\title{Perform Smoothing and / or m-th Derivative}
\usage{
do_sgolay(dataset, p = 2, n = 21, m = 0, exportModel = FALSE)
}
\arguments{
\item{dataset}{An object of class 'aquap_data" as produced by 
\code{\link{gfd}} or as can be extracted from a 'cube' via 
\code{\link{getcd}}.}

\item{p}{Numeric length one, the order of the filter. Default = 2.}

\item{n}{Numeric length one, the filter length, must be odd. Default = 21.}

\item{m}{Numeric length one, Return the m-th derivative of the filter 
coefficients. Default = 0.}

\item{exportModel}{Logical. If a possible model should be stored in the set,
(leave at \code{FALSE}).}
}
\value{
Returns the dataset with the NIR-data smoothed or transformed.
}
\description{
Performs a smoothing and / or the calculation of the m-th 
derivative on the given dataset using the Savitzky-Golay filter.
}
\details{
The underlying function is \code{\link[signal]{sgolayfilt}}.
}
\examples{
\dontrun{
fd <- gfd() # get the full dataset
fd <- selectWls(fd, 1100, 1800)
fd_avg <- do_avg(fd, n=71) # apply strong smoothing
plot(fd - fd_avg, pg.where="", pg.main="| smoothed subtracted")
}
}
\seealso{
Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_addNoise}}, \code{\link{do_avg}},
  \code{\link{do_detrend}}, \code{\link{do_emsc}},
  \code{\link{do_gapDer}}, \code{\link{do_msc}},
  \code{\link{do_resampleNIR}}, \code{\link{dpt_modules}},
  \code{\link{selectWls}}, \code{\link{ssc}}

Other dpt modules documentation: \code{\link{do_detrend}},
  \code{\link{do_emsc}}, \code{\link{do_gapDer}},
  \code{\link{do_msc}}, \code{\link{do_snv}},
  \code{\link{dpt_modules}}
}
