% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_importData.r
\name{export_ap2_ToXlsx}
\alias{export_ap2_ToXlsx}
\title{Export Data to xlsx}
\usage{
export_ap2_ToXlsx(dataset, md = getmd(), onlyNIR = FALSE, rowns = TRUE)
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{md}{List. The object with the metadat of the experiment. 
The default is to get the metadata file via  \code{\link{getmd}}.}

\item{onlyNIR}{Logical. If only the NIR data should be exported to xlsx. If 
left at the default \code{FALSE}, the header as well as the NIR data will 
be exported into a single worksheet.}

\item{rowns}{Logical. If rownames should be exported as well. Defaults to 
\code{TRUE}.}
}
\value{
Invisible \code{TRUE} for a successful, invisible \code{FALSE} for 
an unsuccessful generation of the xlsx file. Used for its side effect, i.e. 
to export a dataset to xlsx.
}
\description{
Export a dataset to xlsx. The header and the NIR data are 
being exported into a single worksheet. In an additional worksheet called 
the name of the experiment with \code{_meta} as suffix, the number of columns 
in the header is denoted, as well as if the first column is the rownames.
In the third column called \code{ncpwl} the number of characters
before the actual wavelength is denoted.
}
\details{
By providing \code{expName="xxx"} in the \code{getmd} function at 
parameter \code{md} it is possible to save the generated xlsx file under 
a different name. See examples.
}
\examples{
\dontrun{
fd <- gfd()
export_ap2_ToXlsx(fd)
export_ap2_ToXlsx(fd, md=getmd(expName="otherName"))
}
}
\seealso{
Other Helper Functions: 
\code{\link{ap2_settings_setup}()},
\code{\link{ap2dme}()},
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{getstn}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()},
\code{\link{updateSettings}()}
}
\concept{Helper Functions}
