% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_importData.r
\name{export_header_toXls}
\alias{export_header_toXls}
\title{Export Header to Xls}
\usage{
export_header_toXls(
  dataset,
  md = getmd(),
  asSlIn = FALSE,
  rowns = TRUE,
  cutCCC = TRUE
)
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{md}{List. The object with the metadat of the experiment. 
The default is to get the metadata file via  \code{\link{getmd}}.}

\item{asSlIn}{Logical, if the created xlsx file should be named ending in 
\code{-in}, thus being ready to reimport as sample list file. Defaults to 
\code{FALSE}. **Warning!** If \code{asSlIn} is set to \code{TRUE}, 
any existing sample list with the same name will be overwritten without 
warning!}

\item{rowns}{Logical. If rownames should be exported as well. Defaults to 
\code{TRUE}.}

\item{cutCCC}{Logical. If the numeric columns possibly copied as class 
variables during rawdata import should be omitted. Defaults to \code{TRUE}.
(cutCCC: 'cut copied C-Cols')}
}
\value{
\code{TRUE} if the operation was successful, \code{FALSE} if not. 
Used for its side effects, i.e. the header exported to xls in the \code{sl-in} 
folder.
}
\description{
Exports only the header, i.e. the class- and numerical variables, 
to an xls-file in the folder \code{sl-in}. The filename is the experiment name, 
followed by the suffix \code{_fromHeader} if parameter \code{asSlIn} is 
\code{FALSE}. If it is \code{TRUE}, a possibly existing sampleList-in file 
with the same name will be overwritten.
}
\section{Warning}{

If \code{asSlIn} is set to \code{TRUE}, any existing sample list with the 
same name will be overwritten without warning!
}

\seealso{
Other Helper Functions: 
\code{\link{ap2_settings_setup}()},
\code{\link{ap2dme}()},
\code{\link{export_ap2_ToXlsx}()},
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{getstn}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()},
\code{\link{updateSettings}()}
}
\concept{Helper Functions}
