% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_SampleList.r
\name{exportSampleList}
\alias{exportSampleList}
\alias{esl}
\title{Create and Export Sample Lists}
\usage{
exportSampleList(
  md = getmd(),
  form = "def",
  rnd = TRUE,
  showFirstRows = TRUE,
  timeEstimate = FALSE
)

esl(
  md = getmd(),
  form = "def",
  rnd = TRUE,
  showFirstRows = TRUE,
  timeEstimate = FALSE
)
}
\arguments{
\item{md}{An object of class \code{aquap_md} containing the metadata of the 
experiment. Defaults to \code{getmd()}, what is calling the default filename 
for the metadata file. See \code{\link{getmd}} and \code{\link{metadata_file}}.}

\item{form}{Character, can be 'txt' to export the sample list in a tab 
delimited text file, 'xls' to export as an Excel file, or if left at the 
default 'def' is reading in the value from the settings.r file (parameter 
\code{p_sampleListExportFormat}).}

\item{rnd}{Logical, if the sample list should be randomized or not, defaults 
to TRUE. (Having a non-randomized sample list can be interesting to check the 
correctness of the sample list when designing the experiment.}

\item{showFirstRows}{Logical. If the first rows of the sample list should be 
displayed.}

\item{timeEstimate}{Logical. If time estimates should be displayed.}
}
\value{
An (invisible) data frame with a (randomized) sample list resp. this 
list saved to a file.
}
\description{
Creates and exports the randomized sample list to file.
}
\details{
Possible formats to export are tab delimited text or Excel. 
When the appropriate values at the bottom of the settings.r file are provided, 
an approximate time estimate of how long it will take to work through the 
sample list is given.
}
\section{Note}{
 You can provide your own values for how many seconds you 
need for a single scan etc. at the bottom of the settings.r file to make the 
time estimate valid.
}

\examples{
\dontrun{
metadata <- getmd()
sl <- exportSampleList(metadata)
sl <- esl() 	# is the same as above
sl <- esl(md=getmd("myMetadataFile.r"), form="txt", showFirstRows=FALSE)
# this would load the metadata from the file "myMetadataFile.r" and generate 
# a sample list based on these metadata. This would be a rare case, as usually 
# in an experiment home-folder you will have probably only one metadata file 
# but several analysis procedure files.
}
}
\seealso{
Other Core functions: 
\code{\link{gdmm}()},
\code{\link{getFullData}()},
\code{\link{plot,aquap_cube,missing-method}},
\code{\link{plot,aquap_data,missing-method}}
}
\concept{Core functions}
\concept{Import-Export}
