% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_generateDatasets.r
\name{gdmm}
\alias{gdmm}
\title{*** Generate Datasets and make Models ***}
\usage{
gdmm(dataset, ap = getap(), md = getmd())
}
\arguments{
\item{dataset}{An object of class 'aquap_data' as generated by \code{\link{gfd}}}

\item{ap}{The analysis procedure, an object of class 'aquap_ap'}

\item{md}{The metadata, an object of class 'aquap_md'}
}
\value{
An object of class \code{\link{aquap_cube}} containing all the 
statistical models / calculations that were performed on the split-variations 
of the dataset.
}
\description{
Generate several datasets by splitting up the original dataset 
according to the variables and values as specified in the analysis procedure 
in the 'split dataset' section and then calculate the models as specified 
in the 'statistics' section on all of the  datasets. By providing additional 
arguments to the function \code{\link{getap}} (what is the default way to get 
the analysis procedure 'ap') you can override any value in the analysis 
procedure. Please see examples and \code{\link{getap}} and  
\code{\link{anproc_file}} for further information.
}
\details{
Split-combinations that yield no result will automatically be omitted.
It is recommended to first  make the analysis procedure file as complete and 
accurate as possible, and then to override only a few parameters if necessary.
}
\seealso{
\code{\link{getap}}, \code{\link{getmd}}

Other Core functions: \code{\link{exportSampleList}},
  \code{\link{getFullData}}, \code{\link{plot}}
}

