% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{genFolderStr}
\alias{genFolderStr}
\title{Generate Folder Structure}
\usage{
genFolderStr(where = getwd())
}
\arguments{
\item{where}{Character length one, the root folder of a single experiment.
Defaults to \code{getwd()}.}
}
\value{
Used for its side effects, i.e. folders get created.
}
\description{
Generate the required folder structure for a new experiment in 
the current working directory.
}
\details{
\code{genFolderStr} will generate all the required folders in the 
directory specified at \code{where} that 'aquap2' needs to work properly. 
Templates for metadata and analysis procedure will be copied into the 
metadata-folder.
}
\seealso{
\code{\link{settings_file}}, \code{\link{noise_procedures}}, 
\code{\link{genNoiseRecExp}}

Other Helper Functions: 
\code{\link{ap2_settings_setup}()},
\code{\link{ap2dme}()},
\code{\link{export_ap2_ToXlsx}()},
\code{\link{export_header_toXls}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{getstn}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()},
\code{\link{updateSettings}()}
}
\concept{Helper Functions}
