% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aqg.r
\name{genTempCalibExp}
\alias{genTempCalibExp}
\title{Generate temperature recording experiment}
\usage{
genTempCalibExp(Tcenter = NULL, Tdelta = 5, stepBy = 1, repls = 4)
}
\arguments{
\item{Tcenter}{Numeric length one. The temperature at which usually the 
measurements are performed. The final temperature will range from 
Tcenter-Tdelta to Tcenter+Tdelta, in steps given by argument 'stepsBy'.}

\item{Tdelta}{Numeric length one, defaults to 5. The temperature range below 
and above 'Tcenter'.}

\item{stepBy}{Numeric length one, defaults to 1. The temperature step between 
each single temperature in the range from Tcenter-Tdelta to Tcenter+Tdelta.}

\item{repls}{Numeric length one. How many replicates of each single temperature 
to record. Defaults to 4.}
}
\description{
Generate the folder structure for a new experiment and populate 
it with the metadata suggested for recording then the temperature 
calibration-spectra used e.g. in the aquagram (see argument \code{aqg.TCalib} 
and \code{aqg.Texp} in \code{\link{calc_aqg_args}}).
}
\details{
This generates the folder structure for a standard experiment and is 
adapting the metadata to record spectra at various temperatures in each 3 
consecutive scans. For a possible workflow please see examples.
}
\section{Important}{
 When exporting the sample list via \code{\link{esl}}, make 
sure to export it \strong{non randomized} - please see examples.
}

\section{Warning}{
 Do not change the name of the columns in the sample list 
before importing the dataset; if the numerical column \code{smpTemp} is not 
present, the temperature calibration data can not be used.
}

\examples{
\dontrun{
genTempCalibExp(Tcenter=30) # generate the folder structure in the current 
working directory
esl(rnd=FALSE) # export a *non* randomized sample list
#### now record the temperature-spectra #### (move sample list to folder 'sl_in')
gfd <- gfd() # imports temperature raw-data and saves an R-data file in the 
# R-data folder, from where you take it and move it into your 
# AQUAP2SH folder
}
}
\seealso{
\code{\link{tempCalib_procedures}}, \code{\link{genFolderStr}}, 
\code{\link{genNoiseRecExp}}

Other Temperature procedures: \code{\link{tempCalib_procedures}}

Other Helper Functions: \code{\link{genFolderStr}},
  \code{\link{genNoiseRecExp}}, \code{\link{getcd}},
  \code{\link{getcm}}, \code{\link{instAquap2Examples}},
  \code{\link{printStdColnames}}, \code{\link{reColor}},
  \code{\link{saveAQdata}}, \code{\link{siWlg}},
  \code{\link{updateAquap2}}, \code{\link{updateSettings}}
}
\concept{Helper Functions}
\concept{Temperature calibration procedures}
\concept{Temperature procedures}
