% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prep_metadata.r
\name{getap}
\alias{getap}
\title{Get Analysis Procedure}
\usage{
getap(fn = "def")
}
\arguments{
\item{fn}{Character length one. If left at 'def', the default filename for an
analysis procedure file as specified in the settings (factory default is
"anproc.r") is read in. Provide any other valid name of an analysis procedure
file to load it. (Do not forget the '.r' at the end.)}
}
\value{
A list with the analysis procedure.
}
\description{
Read in the analysis procedure from the default or a custom
analysis procedure file located in the metadata-folder.
}
\details{
The name of the default analysis procedure file can be specified in
the settings.
}
\examples{
\dontrun{
ap <- getap()
ap <- getap("myFile.r")
}
}
\seealso{
\code{\link{anproc_file}}, \code{\link{getmd}}
}

