% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{getcd}
\alias{getcd}
\title{Get a single dataset from the 'cube'}
\usage{
getcd(cube, index)
}
\arguments{
\item{cube}{An object as created by \code{\link{gdmm}}.}

\item{index}{The index of the dataset to be obtained. See the leftmost
rowname of the 'cube' object.}
}
\value{
A standard dataset as e.g. produced by the function \code{\link{gfd}}.
}
\description{
Get a single dataset, referenced by its index, from the 'cube' 
object, i.e. the object generated by \code{\link{gdmm}}.
}
\details{
A valid index from the available range (see the row-names of the 
'cube-object') has to be provided.
}
\examples{
\dontrun{
dataset <- gfd()
cube <- gdmm(dataset)
# assumes that in the analysis procedure we have a split variable defined.
dataset_3 <- getcd(cube, 3)
dataset_3
str(dataset_3)
}
}
\seealso{
\code{\link{do_msc}}, \code{\link{do_avg}}

Other Extract Elements: \code{\link{getHeader}},
  \code{\link{getWavelengths,aquap_data-method}},
  \code{\link{getcm}}, \code{\link{siWl}}

Other Helper Functions: \code{\link{genFolderStr}},
  \code{\link{genNoiseRecExp}},
  \code{\link{genTempCalibExp}}, \code{\link{getcm}},
  \code{\link{instAquap2Examples}},
  \code{\link{printStdColnames}}, \code{\link{saveAQdata}},
  \code{\link{updateAquap}}, \code{\link{updateSettings}}
}

