% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{getcm}
\alias{getcm}
\title{Get a single model from the 'cube'}
\usage{
getcm(cube, index, what = "pca")
}
\arguments{
\item{cube}{An object as created by \code{\link{gdmm}}.}

\item{index}{The index of the dataset to be obtained. See the leftmost
rowname of the 'cube' object.}

\item{what}{Character length one, see details.}
}
\description{
Get a single model, referenced by its index and specified by the
argument \code{what} from the 'cube' object, i.e. the object generated by 
 \code{\link{gdmm}}.
}
\details{
The default for \code{what} is "pca. Provide any of 
'pca', 'simca', 'plsr', 'aquagr' to obtain its corresponding model. Possible 
values for \code{what} are:
\describe{
\item{pca}{Gives the back the pca-model of the selected 
dataset from the cube}
 \item{simca}{Gives back the simca-model of the selected 
 dataset from the cube}
 \item{plsr}{Gives back the plsr-model of the selected 
 dataset from the cube}
}
}
\section{Extracting Vectors}{

To extract e.g. two loading vectors from a pca-model, you could use a code like 
\code{loadingVectors <- pcaModel$loadings$[, c(1,2)]} to extract the first two 
loading vectors from the model - see examples.
}

\examples{
\dontrun{
fd <- gfd()
cube <- gdmm(fd)
# assumes that in the analysis procedure we have a split variable defined.
fd_3_pca <- getcm(cube, 3)
str(fd_3_pca)
ld12 <- fd_3_pca$model$loadings[, c(1,2)] # extract the first two loadings
ld24 <- fd_3_pca$model$loadings[, c(2,4)] # extract loadings 2 and 4
fd_2_pls <- getcm(cube, 2, "pls")
str(fd_2_pls)
}
}
\seealso{
\code{\link{do_emsc}}, \code{\link{dpt_modules}}

Other Helper Functions: 
\code{\link{ap2_settings_setup}()},
\code{\link{ap2dme}()},
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getstn}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()},
\code{\link{updateSettings}()}

Other Extract Elements: 
\code{\link{getHeader}()},
\code{\link{getWavelengths,aquap_data-method}},
\code{\link{getcd}()},
\code{\link{siWlg}()},
\code{\link{siWl}()}
}
\concept{Extract Elements}
\concept{Helper Functions}
