% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{getCubeModel}
\alias{getCubeModel}
\title{Get a single model from the 'cube'}
\usage{
getCubeModel(cube, index, what = "pca")
}
\arguments{
\item{cube}{An object as created by \code{\link{gdmm}}.}

\item{index}{The index of the dataset to be obtained. See the leftmost
rowname of the 'cube' object.}

\item{what}{Character length one, see details.}
}
\description{
Get a single model, referenced by its index and specified by the
argument \code{what} from the 'cube' object, i.e. the object generated by 
 \code{\link{gdmm}}.
}
\details{
The default for \code{what} is "pca. Provide any of 
'pca', 'sim', 'pls' to obtain its corresponding model. Possible 
values for \code{what} are:
\describe{
\item{pca}{Gives the back the pca-model of the selected 
dataset from the cube}
 \item{sim}{Gives back the simca-model of the selected 
 dataset from the cube}
 \item{pls}{Gives back the plsr-model of the selected 
 dataset from the cube}
}
}
\examples{
\dontrun{
fd <- gfd()
cube <- gdmm(fd)
# assumes that in the analysis procedure we have a split variable defined.
fd_3_pca <- getCubeModel(cube, 3)
str(fd_3_pca)
fd_2_pls <- getCubeModel(cube, 2, "pls")
str(fd_2_pls)
}
}
\seealso{
\code{\link{do_emsc}}

Other Extract Elements: \code{\link{getCubeDataset}},
  \code{\link{getHeader}}
}

