% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prep_metadata.r
\name{getmd}
\alias{getmd}
\title{Get Metadata}
\usage{
getmd(fn = "def")
}
\arguments{
\item{fn}{Character length one. If left at 'def', the default filename for
a metadata-file as specified in the settings (what is "metadata.r") is read
in. Provide any other valid name of a metadata-file to load it. (Do not forget
the '.r' at the end.)}
}
\value{
A list with all the metadata of the experiment.
}
\description{
Read in the metadata from the default or a custom metadata file
located in the metadata-folder.
}
\details{
The name of the default metadata-file can be specified in the settings..
}
\examples{
\dontrun{
md <- getmd()
md <- getmd("myFile.r")
}
}
\seealso{
\code{\link{metadata_file}}
}

