% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniset_functions.R
\name{getstn}
\alias{getstn}
\title{Get settings object}
\usage{
getstn()
}
\value{
An (invisible) list with the key=value pairs from the settings.R 
file.
}
\description{
Is reading in the key=value pairs from the setting.R file.
}
\seealso{
Other Helper Functions: 
\code{\link{ap2_settings_setup}()},
\code{\link{ap2dme}()},
\code{\link{export_ap2_ToXlsx}()},
\code{\link{export_header_toXls}()},
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()},
\code{\link{updateSettings}()}
}
\concept{Helper Functions}
