% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_importData.r
\name{imp_searchAskColumns}
\alias{imp_searchAskColumns}
\title{Search and Ask for Column Representation}
\usage{
imp_searchAskColumns(allC_var, allY_var, slType = .ap2$.slType)
}
\arguments{
\item{allC_var}{A data frame containing only class variables.}

\item{allY_var}{A data frame containing only numerical variables.}

\item{slType}{What type of sample list is used during the import. If no 
additional sample list is used (slType = 'NULL'), the function stops if no 
column for the nr. of consec. scan is present or gets assigned. You can get 
the current slType with '.ap2$.slType' - see example.}
}
\value{
All the list elements needed in the \code{\link{custom_import}} function 
except 'timestamp', 'info' and 'NIR get assigned in the environment from where 
this function was called.
}
\description{
This function can be used in a custom import function. All 
required list elements except 'timestamp', 'info' and 'NIR' will be assigned in 
the environment from where this function was called.
}
\details{
This function checks if any of the in the settings.r file defined 
standard column names are present in the provided class-variables and numerical 
variables. If no exact match is found, the function asks which one of the 
provided columns is representing each of the standard columns. Those columns 
that do not get assigned to one of the standard columns will be imported under 
their respective name with the corresponding prefix for class- or numerical 
variables.
}
\examples{
\dontrun{
# this could be the content of the .r file for defining a custom function, 
# in this example for an import from an Excel-file.
fileExtension <- ".xlsx"
##
spectralImport <- function(dataFile) {
   require(openxlsx)
   import <- openxlsx::read.xlsx(dataFile)
   #
   allY_var <- import[,c(1,2,10,11,6)]
   allC_var <- import[, c(3,4,8,9,5,7)] 
   info <- list(nCharPrevWl = 2)
   NIR <- as.matrix(import[, 12:18])
   rownames(NIR) <- paste("S", 1:nrow(NIR), sep="")
   timestamp <- NULL
   imp_searchAskColumns(allC_var, allY_var, .ap2$.slType) 
   # assigns all list elements except timestamp, info and NIR
   #
   return(list(sampleNr=sampleNr, conSNr=conSNr, timePoints=timePoints, 
   ecrm=ecrm, repl=repl, group=group, temp=temp, relHum=relHum, C_cols=C_cols, 
   Y_cols=Y_cols, timestamp=timestamp, info=info, NIR=NIR))
   #
 } # EOF
}
}
\seealso{
Other Development Functions: \code{\link{custom_TRH}},
  \code{\link{custom_import}}, \code{\link{readHeader}},
  \code{\link{readSpectra}}
}
