% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_plsr.r
\name{indepPlsPred}
\alias{indepPlsPred}
\title{Independent plsr prediction}
\usage{
indepPlsPred(indepDataset, numVar, cube)
}
\arguments{
\item{indepDataset}{The dataset containing the independent data. An object 
of class 'aquap_data' as produced by \code{\link{gfd}}.}

\item{numVar}{Character length one. The name of a numeric variable in the 
independent dataset.}

\item{cube}{An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.}
}
\description{
Use independent data for predictions in the pls-models within 
the cube
}
\details{
XXX
}
\examples{
\dontrun{
XXX	
}
}

