% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{instAquap2Examples}
\alias{instAquap2Examples}
\title{Install Examples}
\usage{
instAquap2Examples()
}
\description{
Install a single experiment-home folder containing various 
examples.
}
\details{
The example folder will be installed in the directory as specified 
in the .Renviron file. (see \code{\link{updateSettings}})
}
\seealso{
Other Helper Functions: \code{\link{genFolderStr}},
  \code{\link{genNoiseRecExp}},
  \code{\link{genTempCalibExp}}, \code{\link{getcd}},
  \code{\link{getcm}}, \code{\link{printStdColnames}},
  \code{\link{saveAQdata}}, \code{\link{updateAquap2}},
  \code{\link{updateSettings}}
}

