% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen_general.r
\name{loadAquapDatapackage}
\alias{loadAquapDatapackage}
\title{Load the aquap2 data package.}
\usage{
loadAquapDatapackage(branch = "master")
}
\arguments{
\item{branch}{Character, the name of the branch to downlaod. Defaults to
"master".}
}
\description{
Download and install the latest version of package 'aquapData'
 from its github repository.
	Package 'aquapData' contains the data and examples used in package 'aquap2'.
}
\details{


Always downloads and installs the latest available version, also
 if the same up-to-date version is already installed.
}
\examples{
\dontrun{
 loadAquapDatapackage()
 }
}
\seealso{
Other Helper Functions: \code{\link{genFolderStr}};
  \code{\link{loadAQdata}}, \code{\link{saveAQdata}};
  \code{\link{updateAquap}}; \code{\link{updateSettings}}
}

