% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.r, R/mod_modifyDatasets.r
\name{mergeDatasets,aquap_data,aquap_data,missing-method}
\alias{mergeDatasets,aquap_data,aquap_data,missing-method}
\alias{mergeDatasets,aquap_data,aquap_data,aquap_mergeLabels-method}
\alias{mergeDatasets,list,missing,missing-method}
\alias{mergeDatasets,list,missing,aquap_mergeLabels-method}
\alias{mergeDatasets}
\title{Merge Datasets}
\usage{
\S4method{mergeDatasets}{aquap_data,aquap_data,missing}(ds1, ds2, mergeLabels = NULL, noMatch = "ask", dol = TRUE)

\S4method{mergeDatasets}{aquap_data,aquap_data,aquap_mergeLabels}(ds1, ds2, mergeLabels, noMatch = "ask", dol = TRUE)

\S4method{mergeDatasets}{list,missing,missing}(ds1, ds2 = NULL, mergeLabels = NULL, noMatch = "ask", dol = TRUE)

\S4method{mergeDatasets}{list,missing,aquap_mergeLabels}(ds1, ds2 = NULL, mergeLabels, noMatch = "ask", dol = TRUE)
}
\arguments{
\item{ds1}{An object of class 'aquap_data' or a list containing any number of 
objects of class 'aquap_data'}

\item{ds2}{An object of class 'aquap_data'.}

\item{mergeLabels}{An object of class 'aquap_mergeLabels' as generated by 
\code{\link{generateMergeLabels}}.}

\item{noMatch}{Character length one. Defines what should happen in the case of 
non-matching header structures, i.e. the column names of the headers of the 
datasets to me merged do not match exactly. The default value is defined in 
the settings.r file (\code{gen_merge_noMatch}). Possible values are: 
\describe{
\item{<%=pv_noMatchChar[1]%>}{The non-matching header-columns in each dataset are 
displayed, and the user is asked interactively what to do, with the three options 
below as possible options.}
\item{delete}{Non-matching header columns are automatically deleted.} # XXX
\item{<%=pv_noMatchChar[3]%>}{Non-matching header columns are automatically renamed 
after the column name in the first fo the datasets to be merged.}
\item{<%=pv_noMatchChar[4]%>}{In case of non-matching header structures, the merging 
process is stopped, with possibly a message being displayed.}
}}

\item{dol}{Logical length one. If outliers should be detected based on the scope 
of the new, merged dataset. Defaults to 'FALSE'.}
}
\value{
An object of class 'aquap_data', with all the single datasets merged
together.
}
\description{
Merge together two or more datasets, and possibly add class- or 
numerical variables to each dataset.
}
\details{
The resulting dataset is void of metadata (object@metadata) and analysis 
procedures.
}
\seealso{
Other dataset modification functions: 
\code{\link{[,aquap_data-method}},
\code{\link{generateMergeLabels}()}
}
\concept{dataset modification functions}
