% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.r, R/plot_all.r
\docType{methods}
\name{plot,aquap_cube,missing-method}
\alias{plot,aquap_cube,missing-method}
\alias{plot_all_modells}
\alias{plot_cube}
\title{*** Plot All Available Models ***}
\usage{
\S4method{plot}{aquap_cube,missing}(x, y, ...)

plot_cube(x, what = "all", ...)
}
\arguments{
\item{x}{An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.}

\item{y}{will be ignored}

\item{...}{Additional parameters to override the values of the graphical 
parameters in the analysis procedure and to override the defaults of the 
arguments in the underlying plotting functions like e.g. \code{\link{plot_pca}}.
Plase see details and the documentation for the other plot functions. 
To specify the source of the analysis procedure and so of the plottig parameters 
use the argument 
\describe{
\item{aps}{Character length one. The default way to obtain the analysis 
procedure. Defaults to "def". Possible values are:
\describe{
\item{"def"}{The default from the settings.r file is taken. (Argument 
\code{gen_plot_anprocSource})}
\item{"cube"}{Take the analysis procedure from within the cube, i.e. the 
analysis procedure that was used when creating the cube via \code{\link{gdmm}}
is used.}
\item{"defFile"}{Use the analysis procedure with the default filename as 
specified in the settings.r file in \code{fn_anProcDefFile}.}
\item{Custom filename}{Provide any valid filename for an analysis procedure to 
use as input for specifying the plotting options.}
}}}}

\item{what}{Character length one. What type of analysis to plot. Possible 
values are 'all', 'pca', 'sim', 'pls', 'aqg'.}
}
\value{
A PDF or graphic device.
}
\description{
Plot all available analysis graphics from the models in the 
cube-object. Function \code{plot} is a convenience function, it basically 
calls  \code{plot_cube}, what is the work-horse for plotting the 
cube-object.
}
\details{
The graphical parameters for plotting are obtained either from the 
standard analysis procedure file, from a custom analysis procedure file, or 
from the analysis procedure embedded in the cube when it was generated via 
\code{\link{gdmm}}. To specify the source of the analysis procedure and so of 
the plottig parameters use the argument \code{aps} in the \code{...} argument;
the default for \code{aps} is "def", meaning that the default value as stored 
in the settings.r file (parameter \code{gen_plot_anprocSource}) is obtained. 
Generally, via the \code{...} argument it is possible to 
override any of the plotting parameters. Please see \code{\link{anproc_file}} 
for a complete listing and e.g. \code{\link{plot_pca_args}} and there the other 
functions of the 'Plot arguments' family for a separate listing of possible 
arguments. Consult the documentation for the other plot functions like e.g. 
\code{\link{plot_pca}} for the arguments in the underlying plot functions 
\strong{not} contained in the analysis procedure --  see examples.
}
\examples{
\dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot(cube, what="all") # the same as above
 plot(cube, what="pca")
 plot(cube, what="pca", pg.main="FooBar") # add "FooBar" to main
 plot(cube, what="pca", aps="foo.r", ld.col=c("green", "blue")) # to load the
 # analysis procedure from a file called "foo.r", and to have the loadings in 
 # the pca drawn in green and blue.
}
}
\seealso{
\code{\link{plot_spectra}}

Other Core functions: \code{\link{exportSampleList}},
  \code{\link{gdmm}}, \code{\link{getFullData}},
  \code{\link{plot,aquap_data,missing-method}}

Other Plot functions: \code{\link{plot,aquap_data,missing-method}},
  \code{\link{plot_aqg}},
  \code{\link{plot_pca,aquap_cube-method}},
  \code{\link{plot_pls,aquap_cube-method}},
  \code{\link{plot_simca,aquap_cube-method}}
}

