% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_aqg.r
\name{plot_aqg_args}
\alias{plot_aqg_args}
\title{Plot Aquagram - Arguments}
\arguments{
\item{do.aqg}{Logical. If used in a plotting function, if Aquagrams should
be plotted.}

\item{fsa}{Fix scale for the Aquagrams.
Logical, numeric or Character. If left at the default logical FALSE,
every single aquagram will be plotted in its own, independent
scale. If a numeric vector length two is provided, all the aquagrams to be
plotted (normal AND bootstrapped ones) will be in the provided range,
no independently scaled aquagrams will be plotted.
If character, the following values are possible:
\code{"both"}: both independently scaled AND automatically calculated
fix-scaled aquagrams will be plotted; \code{"only"}: only the automatically
calculated fix-scaled aquagrams will be plotted. (normal AND bootstrap)}

\item{fss}{Fix scale for the subtraction spectra.
Logical, numeric or character. If left at the default logical FALSE', every
single subtraction-spectra plot will be plotted in its own, independendent
scale. If a numeric vector length two is provided, all the subtraction-spectra
to be plotted (if 'plotSpectra' contains 'subtr', and 'minus' contains a
valid value) will be in the provided range, no independently scaled
subtraction-spectra will be plotted. If character, the following values are
possible: \code{"both"}: both independently scaled AND automatically calculated
fix-scaled spectra will be plotted; \code{"only"}: only the  automatically
calulated fix-scaled subtraction spectra will be plotted.}

\item{pg.where}{Character length one. For plotting PDFs provide "pdf", for
plotting to graphics device provide anything but "pdf".}

\item{pg.main}{Character length one. The additional text on the title of each
single plot.}

\item{pg.sub}{Character length one. The additional text on the subtitle of
each single plot.}

\item{pg.fns}{Character length one. The additional text in the filename of
the pdf.}
}
\description{
The following parameters can be used in the \code{...} argument in
function \code{\link{plot}} and \code{\link{plot_aqg}} to override the values
in the analysis procedure file and so to modify the graphics - see examples.

\code{plot(cube, ...)}

\code{plot_aqg(cube, ...)}
}
\details{
For a list of all parameters that can be used in the \code{...}
argument in \code{\link{getap}} and in the \code{\link{plot}} functions
please see \code{\link{anproc_file}}.
}
\examples{
\dontrun{
dataset <- gfd()
cube <- gdmm(dataset)
plot(cube)
}
}
\seealso{
\code{\link{plot_aqg}}

Other Aquagram documentation: \code{\link{calc_aqg_args}};
  \code{\link{plot_aqg}}

Other Plot arguments: \code{\link{plot_pca_args}};
  \code{\link{plot_pg_args}}; \code{\link{plot_pls_args}};
  \code{\link{plot_sim_args}}
}

