% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_pca.r
\name{plot_pca}
\alias{plot_pca}
\title{PCA - plot}
\usage{
plot_pca(cube, ...)
}
\arguments{
\item{cube}{An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.}

\item{...}{Optional pca plotting parameters to override the values in the
analysis procedure stored in the 'cube' - see below.}

\item{do.pca}{Logical. If used in a plotting function, if PCA score / loading
plots should be plotted.}

\item{pca.colorBy}{NULL or character vector. Which class-variables
should be used for coloring the PCA score plot. Set to NULL for using all
available class variables for coloring.}

\item{pca.what}{Character length one. What element of the PCA analysis to plot.
Possible values are 'both', 'scores', 'loadings'.}

\item{pca.sc}{Numeric length 2. Two PCs to be plotted against each other
in the score plots.}

\item{pca.sc.pairs}{Numeric vector of length >=2, indicating what PCs
to plot in the score pairs plot. Set to NULL for *not* plotting the pairs
plot.}

\item{pca.lo}{Numeric vector of length >=2, indicating what PCs to plot in
the loadingplot.}
}
\value{
A pdf or graphic device.
}
\description{
Plot PCA scoresplots and / or loadings.
}
\details{


The width and height of the resulting pdf can be set in the settings.
}
\seealso{
Other Plot functions: \code{\link{plot}},
  \code{\link{plot,aquap_cube-method}},
  \code{\link{plot_cube}}
}

