% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.r
\name{plot_pca_args}
\alias{plot_pca_args}
\title{Plot PCA - Arguments}
\arguments{
\item{do.pca}{Logical. If used in a plotting function, if PCA score / loading 
plots should be plotted.}

\item{pca.colorBy}{NULL or character vector. Which class-variables 
should be used for coloring the PCA score plot. Set to NULL for using all 
available class variables for coloring.}

\item{pca.elci}{'def' or numeric length one. The confidence interval for the
ellipse to be drawn around groups in score plots. Leave at 'def' to read in 
the default from the settings.r file; provide a numeric length one 
(e.g. 0.95); or set to NULL for not drawing ellipses at all.}

\item{pca.elcolorBy}{Character vector or NULL. The variables to use for 
plotting additional confidence intervall ellipses. Set to NULL for *not* 
drawing additional CI-ellipses. Provide one variable (gets recycled) or a 
vector with equal length as \code{pca.colorBy} to have the additional 
CI-ellipses along these variables.}

\item{pca.what}{Character length one. What element of the PCA analysis to plot.
Possible values are 'both', 'scores', 'loadings'.}

\item{pca.sc}{Numeric length 2. Two PCs to be plotted against each other 
in the score plots.}

\item{pca.sc.pairs}{Numeric vector of length >=2, indicating what PCs 
to plot in the score pairs plot. Set to NULL for *not* plotting the pairs 
plot.}

\item{pca.lo}{Numeric vector of length >=2, indicating what PCs to plot in 
the loadingplot.}

\item{pg.where}{Character length one. For plotting PDFs provide "pdf", for 
plotting to graphics device provide anything but "pdf".}

\item{pg.main}{Character length one. The additional text on the title of each 
single plot.}

\item{pg.sub}{Character length one. The additional text on the subtitle of 
each single plot.}

\item{pg.fns}{Character length one. The additional text in the filename of 
the pdf.}
}
\description{
The following parameters can be used in the \code{...} argument in 
function \code{\link{plot}} and \code{\link{plot_pca}} to override the values 
in the analysis procedure file and so to modify the graphics of score and 
loading plots - see examples.

\code{plot(cube, ...)}

\code{plot_pca(cube, ...)}
}
\details{
For a list of all parameters that can be used in the \code{...} 
argument in \code{\link{getap}} and in the \code{\link{plot}} functions 
please see \code{\link{anproc_file}}.
}
\examples{
\dontrun{
dataset <- gfd()
cube <- gdmm(dataset)
plot(cube, pca.colorBy="C_Group", pca.sc.pairs=NULL) 
# will plot every available model
plot_pca(cube, pca.colorBy="C_Group", pca.sc.pairs=NULL) 
# will plot only the PCA
plot_pca(cube, pca.what="scores")
}
}
\seealso{
\code{\link{plot_pca}}

Other PCA documentation: \code{\link{calc_pca_args}},
  \code{\link{plot_pca}}

Other Plot arguments: \code{\link{plot,aquap_data-method}},
  \code{\link{plot_aqg_args}}, \code{\link{plot_pg_args}},
  \code{\link{plot_pls_args}}, \code{\link{plot_sim_args}}
}

