% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plsr.r
\name{plot_pls}
\alias{plot_pls}
\title{@title Plot PLSR
 @description Plot PLSR XXX.
 @details The width and height of the resulting pdf can be set in the settings.
 @param cube An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.
 @param ... Optional 'pls' plotting parameters to override the values in the 
 analysis procedure stored in the 'cube' - for possible arguments see 
 \code{\link{plot_pls_args}}.
 @return A pdf or graphic device.
 @family Plot functions
 @family PLSR documentation
 @examples
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_pls(cube)
 }
 @export}
\usage{
plot_pls(cube, ...)
}
\description{
@title Plot PLSR
 @description Plot PLSR XXX.
 @details The width and height of the resulting pdf can be set in the settings.
 @param cube An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.
 @param ... Optional 'pls' plotting parameters to override the values in the 
 analysis procedure stored in the 'cube' - for possible arguments see 
 \code{\link{plot_pls_args}}.
 @return A pdf or graphic device.
 @family Plot functions
 @family PLSR documentation
 @examples
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_pls(cube)
 }
 @export
}

