% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plsr.r
\name{plot_pls_args}
\alias{plot_pls_args}
\title{Plot PLSR - Arguments}
\arguments{
\item{do.pls}{Logical. If used in a plotting function, if analysis from PLSR 
models should be plotted.}

\item{pls.colorBy}{NULL or character. What class-variable should be 
used for coloring in the RMSEP and RMSECV plots. Set to NULL for no coloring, 
or provide a character length one with a single column name of a class 
variable that should be used for coloring.}

\item{pg.where}{Character length one. For plotting PDFs provide "pdf", for 
plotting to graphics device provide anything but "pdf".}

\item{pg.main}{Character length one. The additional text on the title of each 
single plot.}

\item{pg.sub}{Character length one. The additional text on the subtitle of 
each single plot.}

\item{pg.fns}{Character length one. The additional text in the filename of 
the pdf.}
}
\description{
The following parameters can be used in the \code{...} argument in 
function \code{\link{plot}} and \code{\link{plot_pls}} to override the values 
in the analysis procedure file and so to modify the graphics - see examples.

\code{plot(cube, ...)}

\code{plot_pls(cube, ...)}
}
\details{
For a list of all parameters that can be used in the \code{...} 
argument in \code{\link{getap}} and in the \code{\link{plot}} functions 
please see \code{\link{anproc_file}}.
}
\examples{
\dontrun{
dataset <- gfd()
cube <- gdmm(dataset)
plot(cube, do.pls=FALSE) # to plot everything availalbe except the plsr-models
plot(cube, pls.colorBy="C_Temp")
plot_pls(cube, pls.colorBy="C_Temp")
}
}
\seealso{
\code{\link{plot_pls}}

Other Plot arguments: \code{\link{plot_aqg_args}},
  \code{\link{plot_pca_args}}, \code{\link{plot_pg_args}},
  \code{\link{plot_sim_args}}

Other PLSR documentation: \code{\link{calc_pls_args}},
  \code{\link{plot_pls}}
}

