% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_plsr.r
\name{plot_pls_indepPred}
\alias{plot_pls_indepPred}
\title{Independent plsr prediction}
\usage{
plot_pls_indepPred(indepDataset, cube, cnv = NULL, inv = NULL,
  aps = "def", ...)
}
\arguments{
\item{indepDataset}{The dataset containing the independent data. An object 
of class 'aquap_data' as produced by \code{\link{gfd}}.}

\item{cube}{An object of class 'aquap_cube' as produced by \code{\link{gdmm}}. 
It is an error to have no plsr models in the cube.}

\item{cnv}{Character vector or NULL. The names of one or more numeric 
variables in the cube on which models have been calculated. Leave at the 
default NULL to use \strong{all} of the numeric variables on which a plsr model 
has been calculated, or provide a character vector with valid variable names 
for a sub-selection. For the selected variables, predictions from the data in 
the independent dataset will be made. If argument \code{inv} is left 
at its default NULL, numerical variables with exactly the same name are looked 
for in the independent dataset and, if present, are used for validating the 
predictions.}

\item{inv}{Character vector or NULL. The names of numeric 
variables in the independent dataset. If left at the default NULL, numerical 
variables in the independent dataset with exactly the same name(s) as specified 
in argument \code{cnv} are looked for and, if present, are used for 
validating the predictions. If a character vector is provided, it has to have 
the same length as the one in \code{cnv}, and those variables will be 
used, in the given sequence, for validating the predictions.}

\item{aps}{Character length one. The default way to obtain the analysis 
procedure. Defaults to "def". Possible values are:
\describe{
\item{"def"}{The default from the settings.r file is taken. (Argument 
\code{gen_plot_anprocSource})}
\item{"cube"}{Take the analysis procedure from within the cube, i.e. the 
analysis procedure that was used when creating the cube via \code{\link{gdmm}}
is used.}
\item{"defFile"}{Use the analysis procedure with the default filename as 
specified in the settings.r file in \code{fn_anProcDefFile}.}
\item{Custom filename}{Provide any valid filename for an analysis procedure to 
use as input for specifying the plotting options.}
}}

\item{...}{For overriding one or more of the plotting parameters, 
please see \code{\link{plot_pls_args}}.}
}
\value{
An (invisible) list containing the numerical results of the 
predictions.
}
\description{
Use independent data for predictions in the pls-models within 
the cube and plot those predictions and, if available, the validation data.
}
\details{
Please see the documentation for the single parameters to see how 
the selection of the available models and the selection of possibly available 
data for validation is handled.
}
\examples{
\dontrun{
XXX	
}
}
\seealso{
Other PLSR documentation: \code{\link{calc_pls_args}},
  \code{\link{plot_pls,aquap_cube-method}},
  \code{\link{plot_pls_args}}
}

