% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simca.r
\name{plot_sim}
\alias{plot_sim}
\title{@title Plot SIMCA
 @description Plot SIMCA interclass distances.
 @details The width and height of the resulting pdf can be set in the settings.
 @param cube An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.
 @param ... Optional 'sim' plotting parameters to override the values in the 
 analysis procedure stored in the 'cube' - for possible arguments see 
 \code{\link{plot_sim_args}}.
 @return A pdf or graphic device.
 @family Plot functions
 @family SIMCA documentation
 @examples
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_sim(cube)
 }
 @export}
\usage{
plot_sim(cube, ...)
}
\description{
@title Plot SIMCA
 @description Plot SIMCA interclass distances.
 @details The width and height of the resulting pdf can be set in the settings.
 @param cube An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.
 @param ... Optional 'sim' plotting parameters to override the values in the 
 analysis procedure stored in the 'cube' - for possible arguments see 
 \code{\link{plot_sim_args}}.
 @return A pdf or graphic device.
 @family Plot functions
 @family SIMCA documentation
 @examples
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_sim(cube)
 }
 @export
}

