% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_simca.r
\name{plot_sim_args}
\alias{plot_sim_args}
\title{Plot SIMCA - Arguments}
\arguments{
\item{do.sim}{Logical. If used in a plotting function, if analysis of SIMCA
models should be plotted.}

\item{pg.where}{Character length one. For plotting PDFs provide "pdf", for
plotting to graphics device provide anything but "pdf".}

\item{pg.main}{Character length one. The additional text on the title of each
single plot.}

\item{pg.sub}{Character length one. The additional text on the subtitle of
each single plot.}

\item{pg.fns}{Character length one. The additional text in the filename of
the pdf.}
}
\description{
The following parameters can be used in the \code{...} argument in
function \code{\link{plot}} and \code{\link{plot_sim}} to override the values
in the analysis procedure file and so to modify the graphics - see examples.

\code{plot(cube, ...)}

\code{plot_sim(cube, ...)}
}
\details{
For a list of all parameters that can be used in the \code{...}
argument in \code{\link{getap}} and in the \code{\link{plot}} functions
please see \code{\link{anproc_file}}.
}
\examples{
\dontrun{
dataset <- gfd()
cube <- gdmm(dataset)
plot(cube, pg.sub="bar") # adds 'bar' to the subtitle of each single plot
plot_sim(cube, pg.sub="bar", pg.fns="_bar") # adds '_bar' to the filename of
# the pdf
}
}
\seealso{
\code{\link{plot_sim}}

Other Plot arguments: \code{\link{plot_aqg_args}};
  \code{\link{plot_pca_args}}; \code{\link{plot_pg_args}};
  \code{\link{plot_pls_args}}

Other SIMCA documentation: \code{\link{calc_sim_args}};
  \code{\link{plot_sim}}
}

