% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.r, R/plot_spectra.r
\docType{methods}
\name{plot,aquap_data-method}
\alias{plot,aquap_data-method}
\alias{plot_spectra}
\title{Plot Raw Spectra}
\usage{
\S4method{plot}{aquap_data}(x, y, ...)

plot_spectra(x, colorBy = NULL, ...)
}
\arguments{
\item{x}{The standard dataset as produced by \code{\link{gfd}}.}

\item{y}{Ignored.}

\item{...}{Optional general plotting options as defined in 
\code{\link{plot_pg_args}}.}

\item{colorBy}{Character vector, possible values are the class variables in 
the dataset.}
}
\description{
Plot the raw spectra contained in the dataset
}
\details{
XXX
}
\examples{
\dontrun{
plot_spectra(dataset)
plot(dataset) # the same as above
plot(dataset, colorBy="C_Group")
plot_spectra(dataset, "C_Group")
plot_spectra(dataset, c("C_Group", "C_Repl"))
plot(dataset, pg.where="") # for plotting to graphic device
plot_spectra(dataset, c("C_Group", "C_Repl"), pg.where="", pg.main="foo")
}
}
\seealso{
plot

Other Plot functions: \code{\link{plot_aqg}},
  \code{\link{plot_pca}}, \code{\link{plot_pls}},
  \code{\link{plot_sim}}, \code{\link{plot}}
}

