% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification.r
\name{plot_SVM}
\alias{plot_SVM}
\title{Plot SVM}
\usage{
plot_SVM()
}
\description{
XXX
}
\details{
XXX
}
\seealso{
Other Plot functions: \code{\link{plot,aquap_cube,missing-method}},
  \code{\link{plot,aquap_data,missing-method}},
  \code{\link{plot_DA}}, \code{\link{plot_NNET}},
  \code{\link{plot_RNF}}, \code{\link{plot_aqg}},
  \code{\link{plot_pca,aquap_cube-method}},
  \code{\link{plot_pls,aquap_cube-method}},
  \code{\link{plot_pls_indepPred}},
  \code{\link{plot_simca,aquap_cube-method}}

Other SVM documentation: \code{\link{calc_SVM_args}},
  \code{\link{plot_SVM_args}}
}
