% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_classification.r
\name{plot_SVM_args}
\alias{plot_SVM_args}
\title{Plot SVM - Arguments}
\arguments{
\item{reserved}{-- No plotting parameter yet defined --}

\item{pg.where}{Character length one. If left at the default 'def', the value 
from the settings.r file is read in (parameter \code{gen_plot_pgWhereDefault}). 
For plotting to PDFs provide "pdf", for plotting to graphics device provide 
anything but "pdf".}

\item{pg.main}{Character length one. The additional text on the title of each 
single plot.}

\item{pg.sub}{Character length one. The additional text on the subtitle of 
each single plot.}

\item{pg.fns}{Character length one. The additional text in the filename of 
the pdf.}
}
\description{
The following parameters can be used in the \code{...} argument 
e.g. in function \code{\link{plot}} and \code{\link{plot_svm}} to override 
the values in the analysis procedure file and so to modify the graphics - 
see examples.
\describe{
\item{\code{plot(cube, ...)}}{ }
\item{ \code{plot_svm(cube, ...)}}{ }
}
}
\details{
For a list of all parameters that can be used in the \code{...} 
argument in \code{\link{getap}} and in the \code{\link{plot}} functions 
please see \code{\link{anproc_file}}.
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{plot_svm}}, \code{\link{siWlg}} for reducing the number of 
wavelengths in a dataset

Other Plot arguments: \code{\link{plot,aquap_data,missing-method}},
  \code{\link{plot_NNET_args}},
  \code{\link{plot_aqg_args}},
  \code{\link{plot_discrimAnalysis_args}},
  \code{\link{plot_pca_args}}, \code{\link{plot_pg_args}},
  \code{\link{plot_pls_args}},
  \code{\link{plot_randomForest_args}},
  \code{\link{plot_sim_args}}

Other SVM documentation: \code{\link{calc_SVM_args}},
  \code{\link{plot_svm,aquap_cube-method}}
}
