% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_plsr.r
\name{pls_outlier_getIndices}
\alias{pls_outlier_getIndices}
\title{PLSR based Outlier Detection}
\usage{
pls_outlier_getIndices(cube, detRange = 0.5)
}
\arguments{
\item{cube}{The standard cube.}

\item{detRange}{The detection range in multiples of interquartile distance.}
}
\value{
A list with logical values marking outliers for each plsr model in 
each cube element.
}
\description{
Obtain the indices of outliers based on a boxplot of plsr data.
}
\details{
Function in development, for internal use only.
}
