% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen_general.r
\name{printStdColnames}
\alias{printStdColnames}
\title{Print standard column names}
\usage{
printStdColnames()
}
\description{
Prints the standard column names as defined in the local
settings.r file to stdout.
}
\seealso{
Other Helper Functions: \code{\link{genFolderStr}};
  \code{\link{loadAQdata}}, \code{\link{saveAQdata}};
  \code{\link{updateAquap}}; \code{\link{updateSettings}}
}

