% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_importData.r
\name{readSpectra}
\alias{readSpectra}
\title{Read Spectra}
\usage{
readSpectra(md = getmd(), filetype = "def", naString = "NA", sh = NULL)
}
\arguments{
\item{md}{List. The object with the metadat of the experiment. 
The default is to get the metadata file via  \code{\link{getmd}}.}

\item{filetype}{Character. The type of the spectral raw data file. 
If a value other than "def" is provided, this is overriding the 
value of "filetype" in the metadata file. Possible values are:
\itemize{
\item \code{def}: Gets the default value from the setings.r file. 
     (variable 'imp_specFileType')
\item \code{vision_NSAS.da}: Import from the .da file generated by the 
Vision-software from a Foss-XDS spectroscope.
\item \code{tabDelim.txt}: Import any tab delimited text file that contains only 
the NIR spectra and *no* additional columns like e.g. time, temperature etc, 
and that has 1 character in front of the wavelengths in the column names of 
the NIR spectra.
\item \code{Pirouette.pir}: Import spectra *and* any class- or numerical variable 
directly from a .pir file. Those column-names in the .pir  file that match the 
standard-column names (\code{\link{printStdColnames}}) as defined in the 
settings.r file will be assigned to those columns automatically.
\item \code{custom@yourFile.R}: It is possible to use a custom import function 
to import **any** type of spectra. Use \code{custom@yourFile.R}, 
with \code{yourFile.R} being an .R-file located in the path 
specified in the .Renviron file. Please refer to \code{\link{custom_import}} 
for further information.
\item \code{xls}: Import raw spectra from a xlsx file. Please see the section
\code{Import from xlsx} below for further details.
\item \code{YunosatoDatFile.dat}: Import raw spectra from a \code{.dat} file
as styled by the Yunosato Aquaphotomics Lab, Japan. Please see the section 
\code{Import from Yunosato .dat} below for further information.
}}

\item{naString}{Character. What to use as 'NA'. Applies only when 'filetype' 
is \code{tabDelim.txt}.}

\item{sh}{Character length one. Manual path to settings home. Can and should 
be left at the default \code{NULL}.}
}
\value{
The spectral data in the format as described in 'Value' in 
\code{\link{custom_import}}
}
\description{
Reads in just the spectra from the provided rawdata file.
}
\details{
This function is mainly for providing a possibility to test custom 
written rawdata import functions. (see \code{\link{custom_import}})
}
\seealso{
\code{\link{getFullData}}

Other Development Functions: 
\code{\link{custom_TRH}},
\code{\link{custom_import}},
\code{\link{imp_searchAskColumns}()},
\code{\link{readHeader}()}
}
\concept{Development Functions}
