% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_SampleList.r
\name{sampleList_multiplyRows}
\alias{sampleList_multiplyRows}
\title{Multiply Rows in Sample List}
\usage{
sampleList_multiplyRows(md = getmd())
}
\arguments{
\item{md}{An object of class \code{aquap_md} containing the metadata of the 
experiment. Defaults to \code{getmd()}, what is calling the default filename 
for the metadata file. See \code{\link{getmd}} and \code{\link{metadata_file}}.}
}
\description{
Reads in the sample list located in the folder 
\code{sampleLists/sl_in} and, if it contains only one row for each sample, 
is multiplying each row by the number of consecutive scans as defined in 
\code{nrConScans} in the metadata file. If a correction value in  is present, 
it is applied. XXX
}
\seealso{
Other Import-Export: 
\code{\link{exportSampleList}()}
}
\concept{Import-Export}
