% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_importData.r
\name{saveAQdata}
\alias{saveAQdata}
\alias{loadAQdata}
\title{Save and load aquap2 datasets}
\usage{
saveAQdata(dataset, md = getmd(), verbose = TRUE)

loadAQdata(md = getmd(), verbose = TRUE)
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{md}{List. The object with the metadat of the experiment. 
The default is to get the metadata file via  \code{\link{getmd}}.}

\item{verbose}{Logical, if messages should be displayed.}
}
\value{
loadData
}
\description{
Save and load the standard aquap2 dataset (class "aquap_data") 
to / from the 'R-data' folder.
}
\details{
From the provided metadata the experiment name is extracted.
\itemize{
 \item saveAQdata The dataset is saved under the same name as the experiment name.
 \item laodAQdata The file having the same name as the experiment name is being 
 loaded.
}
}
\examples{
\dontrun{
saveAQdata(dataset)
loadAQdata()
saveAQdata(dataset, getmd(expName="FooBar")) # save under the name 'FooBar'
}
}
\seealso{
\code{\link{getFullData}}

Other Helper Functions: 
\code{\link{ap2_settings_setup}()},
\code{\link{ap2dme}()},
\code{\link{export_ap2_ToXlsx}()},
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{getstn}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{siWlg}()},
\code{\link{updateSettings}()}
}
\concept{Helper Functions}
