% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube_generateDatasets.r
\name{selectWls}
\alias{selectWls}
\title{Select Wavelengths}
\usage{
selectWls(dataset, from, to)
}
\arguments{
\item{dataset}{XXX}

\item{from}{Numeric length one. XXX}

\item{to}{XXX}
}
\description{
XXX
}
\details{
XXX
}
\seealso{
Other Data pre-preatment functions: \code{\link{ssc}}
}

