% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{siWl}
\alias{siWl}
\title{Isolate single wavelength}
\usage{
siWl(dataset, wl, getMax = FALSE)
}
\arguments{
\item{dataset}{An object of class 'aquap_data' as produced e.g. by 
\code{\link{gfd}}.}

\item{wl}{Numeric length one. The}

\item{getMax}{Logical. Set to 'TRUE' to isolate the wavelength with the 
highest sum of absorbtion values.}
}
\description{
Generate a dataset with a single wavelength.
}
\details{
Provide the wavelength that should remain in the dataset in the 
argument \code{wl}. It is not ncecessary to exactly know the desired 
wavelength -- if there is no direct match with the wavelength, the next best
hit will be taken.
}
\examples{
\dontrun{
fd <- gfd()
dataset_single <- siWl(fd, 1300) # to only leave wavelength 1300
dataset_single2 <- siWl(fd, getMax=TRUE) # to look for the wavelength with 
# highest sum of absorbtion values
}
}
\seealso{
\code{\link{aquap_data-methods}}

Other Extract Elements: \code{\link{getHeader}},
  \code{\link{getcd}}, \code{\link{getcm}}

Other Extract Elements: \code{\link{getHeader}},
  \code{\link{getcd}}, \code{\link{getcm}}
}

