% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{siWlg}
\alias{siWlg}
\title{Reduce Number of Wavelengths}
\usage{
siWlg(dataset, avg = FALSE, wlg = "ot1.12")
}
\arguments{
\item{dataset}{The standard dataset as produced by \code{\link{gfd}}.}

\item{avg}{Logical If, for further reduction of wavelengths, the values of the 
wavelengths in each group (as defined in each row of the matrix) should be 
averaged.}

\item{wlg}{List or character. Provide an integer matrix with two columns and n 
rows to keep the ranges of wavelengths defined in each row - see examples. 
Provide either \code{ot1.12} or \code{ot1.15} to only keep the respective 
wavelengths of the 12 or 15 water matrix coordinates in the first overtone. 
(The definition of the wavelengths is in root of the package aquap2.)}
}
\value{
The transformed dataset.
}
\description{
Reduces the number of wavelengths in a provided dataset, either 
by simply keeping only the wavelengths of the 12 or 15 water matrix coordinates 
in the first overtone (1300nm - 1400nm), or by providing a user-defined list 
containing the wavelengths to be kept.
}
\examples{
\dontrun{
fd <- gfd()
m <- matrix(c(300, 320, 400, 450, 530, 570), ncol=2, byrow=TRUE)
fdc <- siWlg(fd, wlg=m) 
fdc <- siWlg(fd) # using all the 12 wavelength ranges in the first overtone
fdc <- siWlg(fd, TRUE) # averaging within the 12 ranges, resulting in only 12
# wavelengths in the dataset
}
}
\seealso{
\code{\link{selectWls}}, \code{\link{siWl}}

Other Classification Helpers: \code{\link{classification_helpers}}

Other Extract Elements: \code{\link{getHeader}},
  \code{\link{getWavelengths,aquap_data-method}},
  \code{\link{getcd}}, \code{\link{getcm}},
  \code{\link{siWl}}

Other Helper Functions: \code{\link{genFolderStr}},
  \code{\link{genNoiseRecExp}},
  \code{\link{genTempCalibExp}}, \code{\link{getcd}},
  \code{\link{getcm}}, \code{\link{instAquap2Examples}},
  \code{\link{printStdColnames}}, \code{\link{reColor}},
  \code{\link{saveAQdata}}, \code{\link{updateAquap2}},
  \code{\link{updateSettings}}
}
