% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{ssc}
\alias{ssc}
\title{Select Observations}
\usage{
ssc(dataset, criteria, include = TRUE, keepEC = FALSE)
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{criteria}{The selection criteria in the format 
\code{variableName == value}, possibly joined by logical operators.}

\item{include}{Logical. If the observations matching the criteria should be 
included or excluded from the dataset.}

\item{keepEC}{If *all* the environmental control observations should be kept 
in the dataset. Only evaluated if 'include' is TRUE.}
}
\value{
The standard dataset as described in \code{\link{getFullData}}
}
\description{
Create includes or excludes from the dataset by selecting 
from any variable in any logical combination, using the available logical 
operators like e.g. '|' and '&'.
}
\details{
The column names are provided as is, i.e. without quotes, while for 
characters the values have to be enclosed in quotation marks - see examples.
}
\examples{
 \dontrun{
 ds <- ssc(dataset, C_Group=="Control")
 # keeps all the controls
 ds <- ssc(dataset, C_Group!="Control", include=FALSE)
 # the same as above
 
 
 ds <- ssc(dataset, C_Group=="Control" & C_Repl=="R1")
 # keeps only the first replicate of the controls
 ds <- ssc(dataset, C_Group=="Control" | C_Repl=="R1")
 # keeps all the first replicate and all the controls
 
 
 ds <- ssc(dataset, C_Group=="Control" & C_Repl=="R1", keepEC=TRUE)
 # keeps the first replicate of the controls and all the environmental controls
 ds <- ssc(dataset, C_Group=="Control" & C_Repl=="R1", include=FALSE)
 # keeps everything except the first replicate of the controls
 
 
 ds <- ssc(dataset, (C_Group=="Control" | C_Group=="Treatment") & Y_conSNr==1)
 # keeps the first consec. scan of the controls and the treatment group.
 ds <- ssc(dataset, (C_Group=="Control" | C_Group=="MQ") & C_conSNr=="1")
 # keeps the first consec. scan of the controls and the environmental controls
 
 
 ds <- ssc(dataset, Y_Temp==22.5)
 ds <- ssc(dataset, Y_Temp==22.5 & Y_conSNr==1)
 ds <- ssc(dataset, Y_conSNr==1) 
 # keeps only the first consecutive scan
 }
}
\seealso{
Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_avg}}, \code{\link{do_emsc}},
  \code{\link{do_msc}}, \code{\link{do_sgolay}},
  \code{\link{selectWls}}
}

