% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aqg.r
\name{tempCalib_procedures}
\alias{tempCalib_procedures}
\title{Record and use temperature calibration data}
\description{
Record a special temperature dataset and use these data as a 
kind of calibration data for all of the aquagram calculations except the 
'classic' and 'sfc' modes. In other words, you need the temperature dataset 
in order to be able to calculate Aquagrams of the 'auc'modes. It is strongly 
recommended that you do generate the temperature data and so can also use 
the advanced features of these AUC (area under curve) stabilized Aquagrams.
}
\details{
For generating a new experiment with all the necessary defaults to 
record the temperature data, please use \code{\link{genTempCalibExp}} (see 
examples there). After having recorded the spectra in their resp. temperature, 
import the raw data (\code{\link{gfd}}) and move the resulting R-data file 
into your \code{AQUAP2SH} folder. Now in the metadata of any experiment 
(parameter \code{tempCalibFileName}) or in the corresponding parameter in the 
settings file (\code{aqg_tempCalib_Filename}) provide the name of the R-data 
object that you moved into the \code{AQUAP2SH} folder. Whenever now an Aquagram 
is calculated, first and only once per R-session this temperate data file is 
read in and used to calculate the necessary data enabling the calculation of 
'auc' Aquagrams. These temperature-datafile specific objects are stored in 
the environment \code{.ap2} (\code{ls(.ap2)}), starting with the name of the 
temperature data file followed by an '_' underscore.
}
\section{Procedure}{
 The procedure to work with a temperature-data file (or 
more of them of course) and use it to calculate area-under-curve 'auc' 
stabilized Aquagrams is as follows: 
\describe{
\item{Record temperature spectra}{Use the function 
\code{\link{genTempCalibExp}} to generate a folder structure for an experiment, 
export the sample list \strong{non randomized}, then record the 
temperature-spectra. Finally, use \code{\link{gfd}} to import the raw-data 
and create the R-data file (in the folder 'R-data' in the working directory 
of the experiment).}
\item{Move R-data file}{Move the resulting R-data file containing the 
temperature-data from the R-data folder into your \code{AQUAP2SH} folder, 
i.e. the folder also containing e.g. the settings.r file.}
\item{Specify temperature-data file}{In your actual experiment, specify the 
name of the file (residing in the folder \code{AQUAP2SH}) containing the 
temperature-spectra either in the metadata or at the argument \code{tempFile} in the function 
\code{\link{gdmm}} - see examples.}
\item{Choose to calculate an Aquagram}{In your actual experiment, choose to 
actually calculate an Aquagram and specify all the necessary parameters in the 
analysis procedure. You can override the values for the Aquagram-calculations 
via the \code{...} argument in the function \code{\link{getap}} in \code{\link{gdmm}} - please 
see examples and \code{\link{calc_aqg_args}}.}
}
}

\examples{
\dontrun{
fd <- gfd()
cube <- gdmm(fd)
cube <- gdmm(fd, tempFile="def") # to use the default from the settings file, 
# same as above
cube <- gdmm(fd, tempFile="FooBar") # use the temperature-data file 'FooBar'
# residing in the AQUAP2SH folder
cube <- gdmm(fd, getap(do.aqg=TRUE))
cube <- gdmm(fd, getap(do.aqg=TRUE, aqg.bootCI=TRUE))
cube <- gdmm(fd, getap(aqg.mod="aucs.dce"))
cube <- gdmm(fd, getap(aqg.mod="aucs.dce-diff", aqg.minus="C_Cont"), tempFile="FooBar2") 
}
}
\seealso{
\code{\link{genTempCalibExp}}

Other Temperature procedures: \code{\link{genTempCalibExp}}

Other Aquagram documentation: \code{\link{calc_aqg_args}},
  \code{\link{plot_aqg_args}}, \code{\link{plot_aqg}}
}
\concept{Aquagram documentation}
\concept{Temperature procedures}
