% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_transformToAp2.r
\name{ttap}
\alias{ttap}
\title{Transform into aquap2 structure}
\usage{
ttap(obj, expName = "", ncpwl = 1, reduceTRH = FALSE, md = getmd())
}
\arguments{
\item{obj}{The object to be converted}

\item{expName}{An experiment name}

\item{ncpwl}{The numbers of characters before the wavelength.}

\item{reduceTRH}{Logical. If Temp and RelHum values should be reduced.}

\item{md}{The metadata; an object of class 'aquap_md'.}
}
\value{
An object of class 'aquap_data'
}
\description{
Transform a given object into the aquap2 data structure.
}
\details{
Class and Numeric prefixes in the column names are expected
}
\seealso{
Other Development Functions: \code{\link{custom_TRH}},
  \code{\link{custom_import}},
  \code{\link{imp_searchAskColumns}},
  \code{\link{readHeader}}, \code{\link{readSpectra}}
}

