% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{updateAquap}
\alias{updateAquap}
\title{@title Update the aquap2-package.
 @description Download and install the latest version of package 'aquap2' 
 from its github repository
 @details Always downloads and installs the latest available version, also 
 if the same up-to-date version is already installed.
 @param branch Character, the name of the branch to downlaod. Defaults to 
 "master".
 @family Helper Functions
 @examples
 \dontrun{
 updateAquap()
 }
 @export}
\usage{
updateAquap(branch = "master")
}
\description{
@title Update the aquap2-package.
 @description Download and install the latest version of package 'aquap2' 
 from its github repository
 @details Always downloads and installs the latest available version, also 
 if the same up-to-date version is already installed.
 @param branch Character, the name of the branch to downlaod. Defaults to 
 "master".
 @family Helper Functions
 @examples
 \dontrun{
 updateAquap()
 }
 @export
}

