% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{updateAquap2}
\alias{updateAquap2}
\title{Update the aquap2-package.}
\usage{
updateAquap2(force = FALSE)
}
\arguments{
\item{force}{Logical, if the package should be downloaded even if no newer 
version is available. Defaults to \code{FALSE}.}
}
\description{
Download and install the latest version of package 'aquap2' 
from its github repository
}
\examples{
\dontrun{
updateAquap2()
}
}
\seealso{
Other Helper Functions: \code{\link{genFolderStr}},
  \code{\link{genNoiseRecExp}},
  \code{\link{genTempCalibExp}}, \code{\link{getcd}},
  \code{\link{getcm}}, \code{\link{instAquap2Examples}},
  \code{\link{printStdColnames}}, \code{\link{reColor}},
  \code{\link{saveAQdata}}, \code{\link{siWlg}},
  \code{\link{updateSettings}}
}
\concept{Helper Functions}
