% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen_general.r
\name{updateSettings}
\alias{updateSettings}
\title{Update aquap2 settings.}
\usage{
updateSettings(packageName = "aquap2")
}
\arguments{
\item{packageName}{Character, the name of the package where settings
should be updated. Defaults to "aquap2".}
}
\value{
An (invisible) list with the settings \code{stn}
}
\description{
Read in the settings-file in the aquap2-settings home directory
as specified in the .Renviron file
}
\details{
Must be run first after loading the package, and after every change
in the settings.r file to initialize /reload the settings into the
list called 'stn' in the environment '.ap2'.
}
\section{Note}{
 You have to set the path to where you want the settings.r file
to be stored once in your .Renviron file by defining
\code{AQUAP2SH = path/to/any/folder/XX} , with XX being any folder where then the
settings.r file will reside in.
}
\examples{
\dontrun{
updateSettings()
str(.ap2$stn)
ls(.ap2)
}
}
\seealso{
\code{\link{settings_file}}

Other res: \code{\link{settings_file}}
}

