% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_pick+plot.r
\name{adLinesToVector}
\alias{adLinesToVector}
\title{Ad Lines to a Vector Plot}
\value{
Additional graphical elements on a vector plot
}
\description{
Ad various lines / graphical elements to the plot of a 
vector like loading or regression vector.
}
\details{
Leave at the default "def" to read in the respective value from the 
settings.r file. By providing \code{TRUE} or \code{FALSE}, all or none of the 
additional graphical elements are drawn on the vector plot. By providing 
an integer vector in the range [2..5], you can specify which one of the 
graphical elements should be produced:
\itemize{
\item{
 \code{2} Vertical thin lines dropped from a peak to the x-axis. These can 
 help to see in which WAMAC a peak is falling. (See also below at \code{4}.)
}
\item{
 \code{3} The "DelGiudice-areas", three wavelength-ranges in the first overtone 
 corresponding to different water structures / properties. The explicit 
 wavelength ranges are defined in the settings.r file in the parameters 
 \code{dga_dga1} to \code{dga_dga3}, while the other parameters starting with 
 \code{dga_} are governing size and color. The first area is corresponding to 
 the "single" coherent domain, the second to the gaseous phase in between 
 domains, and the third to the meta-domain consisting of coherent single 
 domains. Please note that this is based upon personal communication with 
 Emilio Del Giudice and can not (yet) be referenced or backed by literature 
 or even experimental findings. So, for now these three wavelength ranges are 
 purely hypothecial.If selected for plotting, the three wavelength-ranges will 
 be indicated by slim lines slightly below the x-axis.
}
\item{
 \code{4} The WAMACs, i.e. the so called "water matrix coordinates". If 
 selected for plotting, the location of the WAMACs in the first overtone will 
 be indicated by small, slim boxes in alternating colors directly on the 
 x-axis. The colors themselves and the width in the y-axis of the boxes can be 
 specified in the settings.r file in the parameters starting with \code{wamac_}.
}
\item{
 \code{5} Two single markers indicating the wavelengths for humidity will be 
 plotted on the vector plot. The wavelengths, color, size and line-width are 
 specified in the settings.r file in the parameters starting with 
 \code{hum_}.
}
}
}
\examples{
 \dontrun{
 fd <- gfd()
 cube <- gdmm(fd)
 plot_pca(cube, ld.adLines=FALSE) # no additional lines in the loading plot 
 plot_pca(cube, ld.adLines=c(2, 4) # only vertical lines and WAMACs in the 
 # loading plot
}
}
\seealso{
\code{\link{plot_pca}}, \code{\link{plot_pls}}, 
\code{\link{plot_aqg_args}}
}

