% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/doc_anproc.r
\name{anproc_file}
\alias{anproc_file}
\title{Analysis Procedure File}
\arguments{
\item{spl.var}{NULL or character vector. If NULL, no splitting of the
dataset will be performed. Provide a character vector with the column names
of class variables to split the dataset along these variables.}

\item{spl.wl}{NULL or character vector. If NULL, all in the dataset
available wavelengths will be used. Provide a character vector in the format
"wlFrom-to-wlTo" (e.g. c("1000-to-2000", "1300-to-1600", ...))
to use all previously defined splits in these wavelengths.}

\item{spl.do.smo}{Logical. If smoothing of the NIR data should be performed. The
values for the smoothing operation itself are defined in the settings.}

\item{spl.smo.raw}{Logical. If, should smoothing be performed, the raw
unsmoothed data will be used as well in addition to the smoothed data.}

\item{spl.do.noise}{Logical. If artifical noise should be added to the dataset.}

\item{spl.noise.raw}{If, should the noise-test be performed, the raw data
 will be used as well in addition to the noise-data.}

\item{do.pca}{Logical. If used in \code{gdmm}, if a PCA should be calculated
with a given dataset.}

\item{do.pca}{Logical. If used in a plotting function, if PCA score / loading
plots should be plotted.}

\item{pca.colorBy}{NULL or character vector. Which class-variables
should be used for coloring the PCA score plot. Set to NULL for using all
available class variables for coloring.}

\item{pca.what}{Character length one. What element of the PCA analysis to plot.
Possible values are 'both', 'scores', 'loadings'.}

\item{pca.sc}{Numeric length 2. Two PCs to be plotted against each other
in the score plots.}

\item{pca.sc.pairs}{Numeric vector of length >=2, indicating what PCs
to plot in the score pairs plot. Set to NULL for *not* plotting the pairs
plot.}

\item{pca.lo}{Numeric vector of length >=2, indicating what PCs to plot in
the loadingplot.}

\item{do.sim}{Logical.If SIMCA models of the dataset should be calculated.}

\item{sim.vars}{NULL or character vector. Which variables should be
used to group the data. Set to NULL for using all available class-variables,
or provide a character vector with the column names of class variables to group
the data along those for calculating SIMCA models.}

\item{sim.K}{Numeric length one. The number of components used for calculating
 the SIMCA models. In mode 'robust' leave at '0' for automatic detection of
 optimal number of components. [It is a capital 'K' in the argument.]}

\item{do.pls}{Logical. If PLSR models should be calculated with a given dataset.}

\item{pls.regOn}{NULL or character vector. Which variables should be used
to regress on. Set to NULL for using all numerical variables to regress on, or
provide a character vector with the column names of numerical variables to use
those for regression in the PLSR.}

\item{pls.ncomp}{NULL or integer length one. The number of components used in
PLSR. Set to NULL for automatic detection, or provide an integer to use this
number of components in the PLSR.}

\item{pls.valid}{Character. Which crossvalidation to use. Possible values are
\itemize{
  \item "def" Read in the default value from settings.r
  \item "CV" for a 10-fold crossvalidation
  \item "LOO" for a leave-one-out crossvalidation
  \item XXX something here for using a certain class-variable for grouping XXX
}}

\item{pls.colorBy}{NULL or character. What class-variable should be
 used for coloring in the RMSEP and RMSECV plots. Set to NULL for no coloring,
 or provide a character length one with a single column name of a class
 variable that should be used for coloring.}

\item{do.aqg}{Logical. If Aquagrams should be calculated with a given dataset.}

\item{aqg.vars}{NULL or character vector. Which class variables
should be used for grouping the data for the Aquagram. Provide a character
vector with the column names of class variables for grouping data and generate
an Aquagram for every one of them, or set to NULL if you are providing the
Aquagram class-variable later when calculating and plotting the Aquagram.}

\item{aqg.nrCorr}{Character or Logical. If the number of observations in each
spectral pattern should be corrected (if necessary by random sampling) so
that all the spectral pattern are calculated out from the same number of
observations. If left at the default "def", the default value from the settings
will be used. Provide "TRUE" or "FALSE" to switch number correction manually
on or off.}

\item{aqg.spectra}{Logical or Character. If left at "FALSE" (the default)
no additional spectra are calculated / prepared for plotting. Other possible
values are one or more of:
\itemize{
 \item "raw" for the raw spectra
 \item "avg" for the averaged spectra of the data represented in the aquagram;
 \item "subtr" for subtractions in the averaged spectra (see \code{"minus"} below)
 \item "all" for all of the aforementioned
}}

\item{aqg.minus}{Character. Which of the levels present in the variable
provided in \code{aqg.vars} should be used for subtractions -- the average of
this 'minus' gets subtracted from all the other averages. 'aqg.minus' is used for
the subtractions in the raw spectra as well as for the subtractions within the
Aquagram, should you choose any of the -diff modes.}

\item{aqg.mod}{Character. What mode, what kind of Aquagram should be calculated?
Possible values are: 'classic', 'classic-diff', 'sfc', 'sfc-diff', 'aucs',
'aucs-diff', 'aucs.tn', 'aucs.tn-diff', 'aucs.tn.dce', 'aucs.tn.dce-diff',
'aucs.dce', 'aucs.dce-diff' and 'def' for reading in the default from
settings.r. XXX where do we have the explanation of the mods? XXX}

\item{aqg.TCalib}{Character, numeric or NULL. The default (leave at 'def') can be
set in the settings.  If 'NULL' the complete temperature range of the
calibration data is used for calibration.  Provide a numeric length two
[c(x1, x2)] for manually determining the calibration range.  Provide a
character 'symm@x', with 'x' being the plus and minus delta in temperature
from the temperature of the experiment for having a calibration range from
Texp-x to Texp+x.  The 'Factory' default is 'symm@2'.
Applies to all modes except the 'classic' and 'sfc'  modes.
If, in any of the modes showing percentages, the numbers on the
Aquagram are below 0 or above 100, then the calibration range has to be
extended.}

\item{aqg.Texp}{Numeric length one. The temperature at which the
spectra were taken. The default (leave at 'def') can be set in the settings.}

\item{aqg.bootCI}{Logical. If confidence intervalls for the selected wavelengths
should be calculated within each group (using bootstrap). Leave at 'def' for
getting the default from the settings.}

\item{aqg.R}{Character or numeric. Given aqg.bootCI = TRUE, how many bootstrap
replicates should be performed? Leave at 'def' for choossing the default from
the settings, where the factory-default is "nrow@3" for for 3 x nrow(samples).
By manually providing a character in the form of 'nrow@x' where x is any
number, you can set the factor with which the number of rows get multiplicated,
the result of this multiplication is then used for the number of bootstrap
replicates. By providing a length one numeric you can directly set the number
of bootstrap replicates.}

\item{aqg.smoothN}{Only used in the 'classic' and 'sfc' modes. Numeric length 1.
Must be odd. Smoothing points for the Sav. Golay smoothing that is applied
before making the calculations. Change to NULL or anything not-numeric to
switch off smoothing.}

\item{aqg.selWls}{Only used in the 'classic' and 'sfc' modes.  Numerical vector.
If provided and in the mode "classic", classic-diff", "sfc" and "sfc-diff"
these numbers will be used to determine the coordinates of the aquagram.
Leave at 'def' to use the defaults from the settings file.}

\item{aqg.msc}{Only used in the 'classic' and 'sfc' modes. Logical. If MSC
should be performed.}

\item{aqg.reference}{Only used in the 'classic' and 'sfc' modes. An optional
numerical vector (loadings, etc..) used for MSC.}
}
\description{
The analysis procedure file is used to first split the dataset
 according to the provided values in the 'split dataset' section, and
 then, in the 'statistics' section (starting with \code{do.pca}), to tell
 the system which statistics to apply resp. what models to calculate on those
 datasets. It also contains specific and general plotting options that are used
 by the \code{\link{plot}} function.
 Arguments used to control the split-process, the behaviour of statistics /
 calculations / specific plotting options and the general plotting options
 start with a certain prefix:
 \itemize{
   \item "spl" for all arguments related to the split-process
   \item "pca" for all arguments related to PCA models (except do.pca)
   \item "sim" for all arguments related to SIMCA models (except do.sim)
   \item "pls" for all arguments related to PLSR models (except do.pls)
   \item "aqg" for all arguments related to the calculation of Aquagrams
         (except do.aqg)
   \item "pg" for the general plotting options that are used in each of the
   plotting functions.
 }
 By providing any of the arguments of the analysis procedure file to the
 function \code{\link{getap}}, also when using it inside the function
 \code{\link{gdmm}}, you can override the values in the file with the
 provided values. See examples at \code{\link{gdmm}}.
}
\details{


The default name for the analysis procedure file can be set in
 settings.r. Any other .r file can be loaded by providing a valid .r filename
 to the appropriate argument, e.g. in the function \code{\link{getap}}.
 By providing any of the arguments of the analysis procedure file to the
 function \code{\link{getap}} also when using it inside the function
 \code{\link{gdmm}}, you can override the values in the file with the
 provided values. See examples at \code{\link{gdmm}}.
}
\seealso{
\code{\link{getap}}, \code{\link{gdmm}}

Other fileDocs: \code{\link{metadata_file}};
  \code{\link{settings_file}}
}

