% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aquap2.r
\docType{package}
\name{aquap2}
\alias{aquap2}
\alias{aquap2-package}
\title{Aquaphotomics NIRS Analysis}
\description{
Methods and tools to analyse NIRS data from an Aquaphotomics
perspective. The package can support you in preparing (randomizing) your
experiment, and it will assist you in splitting up your data and analyzing
it using standard multivariate analysis methods and, especially, the Aquagram.
XXX
}
\details{
The first step is to generate the required folder structure for an
experiment via \code{\link{genFolderStr}}. Then XXX
}
\section{Maintainer}{
 Bernhard Pollner <bernhard.pollner@mac.com>
}

\section{Important Functions}{
 List functions here
}
\examples{
\dontrun{
# ad examples here
}
}
\author{
Bernhard Pollner, Zoltan Kovacs
}

