% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquap2.r
\docType{package}
\name{aquap2}
\alias{aquap2}
\title{Multivariate Data Analysis Tools for R including Aquaphotomics Methods}
\description{
Methods and tools for multivariate data analysis in R, with an 
additional focus on specialized Aquaphotomics procedures and methods. 
The package can support you in preparing (randomizing) your 
experiment, and it will assist you in splitting up your data and analyzing 
it using standard multivariate analysis methods and, especially, the Aquagram.
}
\details{
For an introduction and more detailed information please see the package 
vignettes: \code{browseVignettes(package = "aquap2")}
}
\section{Maintainer}{
 Bernhard Pollner <bernhard.pollner@mac.com>
}

\section{Important Functions}{
 \code{\link{esl}}, \code{\link{gfd}} 
\code{\link{gdmm}}, \code{\link{plot_cube}}
}

\examples{
\dontrun{
 dataset <- gfd() # will load or import data
cube <- gdmm(dataset) # split up the dataset and make models
}

}
\author{
Bernhard Pollner, Zoltan Kovacs
}
