% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.r
\docType{methods}
\name{[,aquap_data-method}
\alias{[,aquap_data-method}
\title{'aquap_data' Methods}
\usage{
\S4method{[}{aquap_data}(x, i)
}
\arguments{
\item{x}{An object of class 'aquap_data'}

\item{i}{subsricpting indices for rows}
}
\description{
Available methods for objects of class 'aquap_data'
}
\details{
For subscripting via \code{[}, only values for rownumbers are 
accepted.
}
\note{
\code{drop} is always set to \code{FALSE} for substrictping via \code{[}
}
\examples{
\dontrun{
 dataset <- gfd()
 ds <- dataset[1:5,]
 ds <- dataset[1:5] # the same as above
}
}

