% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_classification.r
\name{calc_discrimAnalysis_args}
\alias{calc_discrimAnalysis_args}
\title{Calculate Discriminant Analysis - Arguments}
\arguments{
\item{do.da}{Logical. If used in \code{getap}, if classification via 
discriminant analysis (\code{\link[MASS]{lda}}, \code{\link[MASS]{qda}}, 
\code{\link[mda]{fda}}, \code{\link[mclust]{MclustDA}}) should be performed 
in the given dataset.}

\item{da.type}{Character vector. The type of discriminant analysis (DA) to 
perform; possible values (one or more) are: 
\code{'lda', 'qda', 'fda', 'mclustda'}:
\itemize{
   \item \code{lda} Linear DA using 
   \code{\link[MASS]{lda}}.
   \item \code{qda} Quadratic DA using 
   \code{\link[MASS]{qda}}.
   \item \code{fda} Flexible DA using 
   \code{\link[mda]{fda}}.
   \item \code{mclustda} DA based on Gaussian finite 
   mixture modeling using \code{\link[mclust]{MclustDA}}.
 }}

\item{da.classOn}{Character vector. One or more class variables to define the 
grouping used for classification.}

\item{da.testCV}{Logical, if the errors of the test-data should be crossvalidated. 
If set to true, CV and testing is repeated in alternating datasets. See below.}

\item{da.percTest}{Numeric length one. The percentage of the dataset that should 
be set aside for testing the models; these data are never seen during training 
and crossvalidation.}

\item{da.cvBootCutoff}{The minimum number of observations (W) that should be 
in the smallest subgroup (as defined by the classification grouping variable) 
*AFTER* the split into \code{da.valid} crossvalidation segments (below). If W 
is equal or higher than \code{da.cvBootCutoff}, the crossvalidation is done 
via splitting the training data in \code{da.valid} (see below) segments, 
otherwise the crossvalidation is done via bootstrap resampling, with the number 
of bootstrap iterations resulting from the multiplication of the number of 
observations in this smallest subgroup (as defined by the classification 
grouping variable) in *all* of the training data with \code{da.cvBootFactor}. 
To never perform the CV of the training data via bootstrap, set the parameter 
\code{cl_gen_neverBootstrapForCV} in the settings.r file to \code{TRUE}. 
An example: With \code{da.cvBootCutoff} 
set to \code{15} and a 8-fold crossvalidation \code{da.valid <- 8}, the 
required minimum number of observations in the smallest subgroup *after* the 
split in 8 segments would be 15, and in all the training data to perform the 
desired 8-fold CV would be (8x15=) 120, in what case then 8 times 15 
observations will form the test data to be projected into models made from 
(120-15=) 105 observations. If there would be less than 120 observations, lets 
say for example, only 100 observations in the smallest group as defined by the 
classification grouping variable, bootstrap resampling with 
\code{da.cvBootFactor * 100} iterations would be performed. In this example, 
if we would also be satisfied with a 5-fold crossvalidation, then we would have 
enough data: 100 / 5 = 20, and with the \code{da.cvBootCutoff} value being 15, 
the 5-fold crossvalidation would be performed.}

\item{da.cvBootFactor}{The factor used to multiply the number of observations 
within the smallest subgroup defined by the classification grouping variable 
with, resulting in the number of iterations of a possible bootstrap 
crossvalidation of the trainign data -- see \code{.cvBootCutoff}.}

\item{da.valid}{The number of segments the training data should be divided 
into in case of a "traditional" crossvalidation of the training data; see above.}

\item{da.pcaRed}{Logical, if variable reduction via PCA should be applied; if 
TRUE, the subsequent classifications are performed on the PCA scores, see
\code{da.pcaNComp} below.}

\item{da.pcaNComp}{Character or integer vector. Provide the character "max" 
to use the maximum number of components (i.e. the number of observations minus 
1), or an integer vector specifying the components resp. their scores to be 
used for DA.}
}
\description{
The following parameters can be used in the \code{...} argument in 
function \code{\link{getap}}, also within function \code{\link{gdmm}}, to 
override the values in the analysis procedure file and so to modify the 
calculation of discriminant analysis models - see examples.
\describe{
\item{\code{getap(...)}}{ }
\item{\code{gdmm(dataset, ap=getap(...))}}{ }
}
}
\details{
For a list of all parameters that can be used in the \code{...} 
argument in \code{\link{getap}} and in the \code{\link{plot}} functions 
please see \code{\link{anproc_file}}.
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{gdmm}}, \code{\link{siWlg}} for reducing the number of 
wavelengths in a dataset

Other Calc. arguments: 
\code{\link{calc_NNET_args}},
\code{\link{calc_SVM_args}},
\code{\link{calc_aqg_args}},
\code{\link{calc_pca_args}},
\code{\link{calc_pls_args}},
\code{\link{calc_randomForest_args}},
\code{\link{calc_sim_args}},
\code{\link{split_dataset}}

Other Classification functions: 
\code{\link{calc_NNET_args}},
\code{\link{calc_SVM_args}},
\code{\link{calc_randomForest_args}},
\code{\link{plot_classifX_indepPred}()}

Other DA documentation: 
\code{\link{plot_da,aquap_cube-method}},
\code{\link{plot_discrimAnalysis_args}}
}
\concept{Calc. arguments}
\concept{Classification functions}
\concept{DA documentation}
