% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_classification.r
\name{classification_helpers}
\alias{classification_helpers}
\title{Classification - General Considerations}
\description{
General remarks regarding the use of the classification functions 
in package aquap2.
\itemize{
\item Except the svm classification, most methods do not work well in fat data 
matrices, so if there are more variables (wavelengths) than observations.
For an easy reduction of the number of wavelengths we provide, especially in 
the Aquaphotomics context, a special function to reduce the number of 
wavelengths in a dataset (as produced by \code{\link{gfd}}): function 
\code{\link{siWlg}} can isolate any custom group of wavelengths from the dataset, 
or, if left at its default, is isolating the ranges of the 12 water matrix 
coordinates within the 1st overtone. For further data reduction, this groups 
can be averaged within as well by setting the second argument in 
\code{\link{siWlg}} to \code{TRUE}.
\item For reducing the number of variables not in the raw data, but on the 
classification side of the algorithm, it is possible to apply all classifiers 
not on the rawdata, but on the PCA scores of the rawdata. This option can be 
activated by setting the respective argument \code{.pcaRed} in the analysis 
procedure to \code{TRUE}. In this case the prediction data for crossvalidation 
as well as the independent test data are projected into the pca-models of the 
training data, and the resulting scores are then used for classification.
\item For both traditional crossvalidation and the bootstrap "crossvalidation", 
consecutive scans of the same sample are always excluded resp. included 
together.
\item A traditional crossvalidation (no bootstrap) is indicated via a "."
after the name of the classificaiton-variable, while the bootstrap
process is indicated via a "`".
}
}
\seealso{
Other Classification Helpers: 
\code{\link{siWlg}()}
}
\concept{Classification Helpers}
