% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_avg}
\alias{do_avg}
\title{Average spectra in a dataset}
\usage{
do_avg(dataset, clv = NULL, inParallel = TRUE)
}
\arguments{
\item{dataset}{An object of class 'aquap_data" as produced by 
\code{\link{gfd}} or as can be extracted from a 'cube' via 
\code{\link{getcd}}.}

\item{clv}{Character vector, one or more valid class-variables defining the 
subsets of the dataset to be averaged.}

\item{inParallel}{Logical, if the averaging of spectra should be done in 
parallel. Defaults to TRUE.}
}
\value{
The transformed dataset.
}
\description{
Calculate the average spectrum of all the given spectra in a 
dataset, or define groups to be averaged by providing one or more 
class-variable.
}
\details{
The total and group-wise average spectrum is returned each in a 
single row with the header-data taken from the first row of the subset of 
the original dataset as defined by the grouping. If parameter \code{clv} is 
left at the default NULL, all the spectra in the dataset will be averaged 
together into a single row.
}
\examples{
\dontrun{
fd <- gfd() # get the full dataset
fd <- selectWls(fd, 1100, 1800)
fd_avg <- do_avg(fd)
fd_avg2 <- do_avg(ssc(fd, C_water=="MQ")) # assumes the existence of column "C_water"
plot(fd_avg, pg.where="", pg.main="| averaged")
plot(fd - fd_avg, pg.where="", pg.main="| avg subtracted")
#######
fd_avg_group <- do_avg(fd, "C_Group") # averaging within groups
fd_avg_time_group <- do_avg(fd, c("C_Time", "C_Group")) # averaging within a 
# subset defined by "C_Time" and "C_Group"
}
}
\seealso{
\code{\link{getcd}}

Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_addNoise}}, \code{\link{do_detrend}},
  \code{\link{do_emsc}}, \code{\link{do_gapDer}},
  \code{\link{do_msc}}, \code{\link{do_resampleNIR}},
  \code{\link{do_sgolay}}, \code{\link{dpt_modules}},
  \code{\link{selectWls}}, \code{\link{ssc}}
}
