% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_msc}
\alias{do_msc}
\title{Perform MSC}
\usage{
do_msc(dataset, ref = NULL)
}
\arguments{
\item{dataset}{An object of class 'aquap_data" as produced by 
\code{\link{gfd}} or as can be extracted from a 'cube' via 
\code{\link{getcd}}.}

\item{ref}{An object of class 'aquap_data' containing a single spectrum,
i.e. a single row (as e.g. produced by \code{\link{do_avg}}, or by 
subscripting via '[]'). If no reference is provided, the average of all spectra 
of the provided dataset is used as a reference for the baseline correction.}
}
\value{
Returns the dataset with the transformed NIR data.
}
\description{
Performs MSC (multiplicative scatter correction) on the 
provided dataset.
}
\details{
If no reference is provided, the average of all spectra of the provided
dataset is used as a reference for the baseline correction. Provide a dataset 
with a single spectrum (as e.g. produced by \code{\link{do_avg}}) to use this 
as a reference for baseline correction. Internally, the function 
\code{\link[pls]{msc}} is used.
}
\examples{
\dontrun{
fd <- gfd() # get the full dataset
fd <- selectWls(fd, 1100, 1800)
fd_msc <- do_msc(fd)
plot(fd - fd_msc, pg.where="", pg.main="| msc subtracted")
fd_msc_ref <- do_msc(fd, ref=fd[4]) # use the 4th row of fd as a reference
plot(fd - fd_msc_ref, pg.where="", pg.main="| 4th row as reference")
fd_msc_mq <- do_msc(fd, ref=do_avg(ssc(fd, C_Group=="MQ"))) # is assuming the 
# existence of a column name "C_Group" and one or more of its values being "MQ"
plot(fd - fd_msc_mq, pg.where="", pg.main="| average of MQ as reference")
}
}
\seealso{
\code{\link{ssc}} for sub-selecting in a datset using a specific 
criterion.

\code{\link{getcd}}

Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_avg}}, \code{\link{do_emsc}},
  \code{\link{do_sgolay}}, \code{\link{selectWls}},
  \code{\link{ssc}}

Other dpt modules documentation: \code{\link{do_emsc}},
  \code{\link{do_sgolay}}, \code{\link{do_snv}},
  \code{\link{dpt_modules}}
}

