% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_sgolay}
\alias{do_sgolay}
\title{Perform Smoothing or m-th Derivative}
\usage{
do_sgolay(dataset, p = 2, n = 21, m = 0)
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{p}{Numeric length 1, the order of the filter.}

\item{n}{Numeric length 1, the filter length, must be odd.}

\item{m}{Numeric length 1, the m-th derivative.}
}
\value{
Returns the dataset with the NIR-data smoothed or transformed.
}
\description{
Performs a smoothing and / or the calculation of the
m-th derivative.
on the given dataset using the Savitzky-Golay filter.
}
\details{
The underlying function is \code{\link[signal]{sgolayfilt}}.
}
\seealso{
Other Data pre-treatment functions: \code{\link{do_msc}};
  \code{\link{selectWls}}; \code{\link{ssc}}
}

