% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coreDataManip.r
\name{do_snv}
\alias{do_snv}
\title{Calculate Standard Normal Variation SNV}
\usage{
do_snv(dataset)
}
\arguments{
\item{dataset}{An object of class 'aquap_data" as produced by 
\code{\link{gfd}} or as can be extracted from a 'cube' via 
\code{\link{getcd}}.}
}
\value{
Returns the dataset with the transformed NIR data.
}
\description{
Calculate the standard normal variation (SNV) by autoscaling the 
transformed NIR-data
}
\examples{
\dontrun{
fd <- gfd() # get the full dataset
fd <- selectWls(fd, 1100, 1800)
fd_snv <- do_snv(fd)
plot(fd - fd_snv, pg.where="", pg.main="| snv subtracted")
}
}
\seealso{
Other dpt modules documentation: \code{\link{do_emsc}},
  \code{\link{do_msc}}, \code{\link{do_sgolay}},
  \code{\link{dpt_modules}}
}

