% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_SampleList.r
\name{exportSampleList}
\alias{exportSampleList}
\alias{esl}
\title{Create and Export Sample Lists}
\usage{
exportSampleList(
  multiplyRows = FALSE,
  rnd = TRUE,
  md = getmd(),
  showFirstRows = TRUE,
  timeEstimate = FALSE
)

esl(
  multiplyRows = FALSE,
  rnd = TRUE,
  md = getmd(),
  showFirstRows = TRUE,
  timeEstimate = FALSE
)
}
\arguments{
\item{multiplyRows}{Logical, defaults to \code{FALSE}. If each row in the 
generated sample list should be multiplied by the number of consecutive scans 
as specified at \code{nrConScans} in the metadata file. 
\itemize{
\item If left at the default \code{FALSE}, there is one row for each sample 
in the sample list. Potential abberations from the planned number of 
consecutive scans can be noted in the inserted column called 
\code{conScanError} by providing \code{±n}, with 'n' being the number the 
actual number of consecutive scans is differing from the intended. 
\item  If \code{multiplyRows} is set to \code{TRUE}, each row in the generated 
sample list is multiplied by the number of consecutive scans as specified at 
\code{nrConScans} in the metadata file. In this case there is no column 
allowing to denote errors, but the user has to manually delete or add rows 
for each sample where the number of consecutive scans is differing from 
from the intended. If rows were added, each cell in the row has to be filled 
accordingly. 
}}

\item{rnd}{Logical, if the sample list should be randomized or not, defaults 
to TRUE. (Having a non-randomized sample list can be interesting to check the 
correctness of the sample list when designing the experiment.}

\item{md}{An object of class \code{aquap_md} containing the metadata of the 
experiment. Defaults to \code{getmd()}, what is calling the default filename 
for the metadata file. See \code{\link{getmd}} and \code{\link{metadata_file}}.}

\item{showFirstRows}{Logical. If the first rows of the sample list should be 
displayed.}

\item{timeEstimate}{Logical. If time estimates should be displayed.}
}
\value{
An (invisible) data frame with a (randomized) sample list resp. this 
list saved to a file.
}
\description{
Creates and exports the randomized sample list to file.
}
\details{
The exclusive format for the sample list is xlsx. 
When the appropriate values at the bottom of the settings.r file are provided, 
an approximate time estimate of how long it will take to work through the 
sample list is given.
}
\section{Note}{
 You can provide your own values for how many seconds you 
need for a single scan etc. at the bottom of the settings.r file to make the 
time estimate valid.
}

\examples{
\dontrun{
metadata <- getmd()
sl <- exportSampleList(metadata)
sl <- esl() 	# is the same as above
sl <- esl(md=getmd("myMetadataFile.r"), form="txt", showFirstRows=FALSE)
# this would load the metadata from the file "myMetadataFile.r" and generate 
# a sample list based on these metadata. This would be a rare case, as usually 
# in an experiment home-folder you will have probably only one metadata file 
# but several analysis procedure files.
}
}
\seealso{
Other Import-Export: 
\code{\link{sampleList_multiplyRows}()}

Other Core functions: 
\code{\link{gdmm}()},
\code{\link{getFullData}()},
\code{\link{plot,aquap_cube,missing-method}},
\code{\link{plot,aquap_data,missing-method}}
}
\concept{Core functions}
\concept{Import-Export}
