% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prep_SampleList.r
\name{exportSampleList}
\alias{esl}
\alias{exportSampleList}
\title{Create and Export Sample Lists}
\usage{
exportSampleList(md = getmd(), form = "def", showFirstRows = TRUE,
  timeEstimate = FALSE)

esl(md = getmd(), form = "def", showFirstRows = TRUE,
  timeEstimate = FALSE)
}
\arguments{
\item{md}{List. An object with the metadata of the experiment. Defaults to
    \code{getmd()}, what is calling the default filename for the metadata file.
See \code{\link{getmd}} and \code{\link{metadata_file}}.}

\item{form}{Character, can be either 'txt' to export the sample list in a tab
delimited text file, or 'xls' to export as an Excel file.}

\item{showFirstRows}{Logical. If the first rows of the sample list should be
displayed.}

\item{timeEstimate}{Logical. If time estimates should be displayed.}
}
\value{
An (invisible) data frame with a randomized sample list resp. this
list saved to  a file.
}
\description{
Creates and exports the randomized sample list to file.
}
\details{
Possible formats to export are XXX. For the time estimates,
	you can fill in your own values at the bottom of the settings.r file.
}
\section{Note}{
 You can provide your own values for how many seconds you
need for a single scan etc. at the bottom of the settings.r file to make the
time estimate valid for you.
}
\examples{
\dontrun{
metadata <- getmd()
sl <- exportSampleList(metadata)
sl <- esl() 	# is the same as above
sl <- esl(md=getmd("myMetadataFile.r"), form="txt", showFirstRows=FALSE)
# this would load the metadata from the file "myMetadataFile.r" and generate
# a sample list based on these metadata. This would be a rare case, as usually
# in an experiment home-folder you will have probably only one metadata file
# but several analysis procedure files.
}
}
\seealso{
Other Core functions: \code{\link{gdmm}};
  \code{\link{getFullData}}, \code{\link{gfd}}
}

