% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube_generateDatasets.r
\name{gdmm}
\alias{gdmm}
\title{Generate Datasets and make Models}
\usage{
gdmm(dataset, md = getmd(), ap = getap())
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{md}{The metadata, an object of class 'aquap_md'}

\item{ap}{The analysis procedure, an object of class 'aquap_ap'}
}
\value{
XXX
}
\description{
Generate several datasets by splitting up the original dataset
according to the variables and values provided in the analysis procedure and
then calculate the models as specified in the analysis procedure on all of the
datasets.
}
\details{
XXX
}
\seealso{
Other Core functions: \code{\link{getFullData}},
  \code{\link{getFullData}}, \code{\link{gfd}},
  \code{\link{gfd}}
}

