% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_generateDatasets.r
\name{gdmm}
\alias{gdmm}
\title{*** Generate Datasets and make Models ***}
\usage{
gdmm(dataset, ap = getap(), noiseFile = "def", tempFile = "def")
}
\arguments{
\item{dataset}{An object of class 'aquap_data' as generated by \code{\link{gfd}}}

\item{ap}{The analysis procedure, an object of class 'aquap_ap'}

\item{noiseFile}{Character length one. The name of the corresponding noise data 
file residing in the AQUAP2SH folder. Leave at the default 'def' to read in the 
value from the settings.r file (parameter \code{noi_noiseDataFilename}), or 
provide the name of an existing R-data file in your specified AQUAP2SH folder 
to use this as source for the noise data. If a value other than "def" is 
provided, this is overriding the value of "noiseFileName" in the metadata file.
Please see \code{\link{noise_procedures}} for more information.}

\item{tempFile}{Character length one. The name of the corresponding temperature 
calibration data file residing in the AQUAP2SH folder. Leave at the default 'def' 
to read in the value from the settings.r file (parameter 
\code{aqg_tempCalib_Filename}), or provide the name of an existing R-data file 
in your specified AQUAP2SH folder to use this as source for the temperature 
calibration data. If a value other than "def" is provided, this is overriding 
the value of "tempCalibFileName" in the metadata file. Please see 
\code{\link{tempCalib_procedures}} for more information.}
}
\value{
An object of class \code{\link{aquap_cube}} containing all the 
statistical models / calculations that were performed on the split-variations 
of the dataset.
}
\description{
Generate several datasets by splitting up the original dataset 
according to the variables and values as specified in the analysis procedure 
in the 'split dataset' section and then calculate the models as specified 
in the 'statistics' section on all of the  datasets. By providing additional 
arguments to the function \code{\link{getap}} (what is the default way to get 
the analysis procedure 'ap') you can override any value in the analysis 
procedure. Please see examples and \code{\link{getap}} and  
\code{\link{anproc_file}} for further information.
}
\details{
Split-combinations that yield no result will automatically be omitted.
It is recommended to first  make the analysis procedure file as complete and 
accurate as possible, and then to override only a few parameters if necessary.
}
\examples{
\dontrun{
dataset <- gfd() # will load or import data
cube <- gdmm(dataset) # split up the dataset and make models, execute the 
# analysis procedure as specified in its .r file
cube <- gdmm(dataset, getap(spl.var="C_Group")) # split the dataset by "C_Group"
cube <- gdmm(dataset, getap(spl.var=c("C_Group", "C_Temp"))) # split the dataset 
# by "C_Group", then by "C_Temp"
cube <- gdmm(dataset, getap(spl.wl="1300-to-1600")) # override 'spl.wl' in the 
# analysis procedure
cube <- gdmm(dataset, getap(aqg.bootCI=FALSE)) # override the value in 
# 'aqg.bootCI' of the analysis procedure with 'FALSE'. 
cube <- gdmm(dataset, getap(do.sim=FALSE, pls.regOn="Y_Temp"))
cube <- gdmm(dataset, getap(spl.do.noise=TRUE), noiseFile="FooBarNoise") # to
# use the noise-data file 'FooBarNoise' instead of the default noise data file
}
}
\seealso{
\code{\link{getap}}, \code{\link{getmd}}

Other Core functions: \code{\link{exportSampleList}},
  \code{\link{getFullData}},
  \code{\link{plot,aquap_cube,missing-method}},
  \code{\link{plot,aquap_data,missing-method}}
}
\concept{Core functions}
