% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube_generateDatasets.r, R/doc_gdmm_classes.r
\name{gdmm}
\alias{gdmm}
\title{*** Generate Datasets and make Models ***}
\usage{
gdmm(dataset, ap = getap(), md = getmd())
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{ap}{The analysis procedure, an object of class 'aquap_ap'}

\item{md}{The metadata, an object of class 'aquap_md'}
}
\value{
An object of class \code{\link{aquap_cube}} containing all the
statistical models / calculations that were performed on the split-variations
of the dataset.
}
\description{
Generate several datasets by splitting up the original dataset
according to the variables and values as specified in the analysis procedure
in the 'split dataset' section and then calculate the models as specified
in the 'statistics' section on all of the  datasets. By providing additional
arguments to the function \code{\link{getap}} (what is the default way to get
the analysis procedure 'ap') you can override any value of the 'statistics'
section in the analysis procedure. Please see examples and \code{\link{getap}}
and  \code{\link{anproc_file}} for further information.
}
\details{
Split-combinations that yield no result will automatically be omitted.
}
\examples{
\dontrun{
dataset <- gfd() # will load or import data
cube <- gdmm(dataset) # split up the dataset and make models, execute the
# analysis procedure as specified in its .r file
cube <- gdmm(dataset, getap(aqg.bootCI=FALSE)) # same as above, but override
the value in 'aqg.bootCI' of the analysis procedure with 'FALSE'.
cube <- gdmm(dataset, getap(do.sim=FALSE, pls.regOn="Y_Temp"))
}
}
\seealso{
\code{\link{getap}}, \code{\link{getmd}}

Other Core functions: \code{\link{esl}}, \code{\link{esl}},
  \code{\link{exportSampleList}};
  \code{\link{getFullData}}, \code{\link{getFullData}},
  \code{\link{gfd}}, \code{\link{gfd}}
}

