% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen_general.r
\name{genFolderStr}
\alias{genFolderStr}
\title{Generate Folder Structure}
\usage{
genFolderStr()
}
\value{
Folders get created in the current working directory.
}
\description{
Generate the required folder structure in the current working
directory.
}
\details{
\code{genFolderStr} will generate all the required folders in the
current working directory that 'aquap2' needs to work properly. Templates
for metadata and analysis procedure will be copied into the metadata-folder.
You can change the defaults for the folder names in the settings file.
}
\seealso{
\code{\link{settings_file}}

Other Helper Functions: \code{\link{instAquap2Examples}};
  \code{\link{loadAQdata}}, \code{\link{saveAQdata}};
  \code{\link{printStdColnames}};
  \code{\link{updateAquap}}; \code{\link{updateSettings}}
}

