% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_noise.r
\name{genNoiseRecExp}
\alias{genNoiseRecExp}
\title{Generate noise recording experiment}
\usage{
genNoiseRecExp()
}
\description{
Generate the folder structure for a new experiment and populate 
it with the metadata suggested for recording then the noise-spectra.
}
\details{
This generates the folder structure for a standard experiment and is 
adapting the metadata to record 12 noise-spectra with each 3 consecutive scans.
For a possible workflow please see examples.
}
\examples{
\dontrun{
genNoiseRecExp() # generate the folder structure in the current working 
# directory
esl() # export the sample list
#### now record the noise-spectra #### (move sample list to folder 'sl_in')
gfd <- gfd() # imports noise raw-data and saves noise-data file in the 
# R-data folder, from where you take it and move it into your 
# AQUAP2SH folder
}
}
\seealso{
\code{\link{genTempCalibExp}}, \code{\link{genFolderStr}}

Other Helper Functions: \code{\link{genFolderStr}},
  \code{\link{genTempCalibExp}}, \code{\link{getcd}},
  \code{\link{getcm}}, \code{\link{instAquap2Examples}},
  \code{\link{printStdColnames}}, \code{\link{saveAQdata}},
  \code{\link{updateAquap}}, \code{\link{updateSettings}}

Other Noise procedures: \code{\link{do_addNoise}},
  \code{\link{noise_procedures}}
}

