% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_metadata.r
\name{getmd}
\alias{getmd}
\title{Get Metadata}
\usage{
getmd(fn = "def", ...)
}
\arguments{
\item{fn}{Character length one.  The filename of the metadata file to load. 
If left at 'def', the default filename for 
a metadata-file as specified in the settings (what is "metadata.r") is read 
in. Provide any other valid name of a metadata-file to load it. (Do not forget 
the '.r' at the end.)}

\item{...}{Provide the argument 'expName' to override the experiment name 
specified in the metadata.r file.}
}
\value{
An (invisible) list with all the metadata of the experiment.
}
\description{
Read in the metadata from the default or a custom metadata file 
located in the metadata-folder. By providing the argument 'expName' to the 
function you can override the experiment name as specified in the metadata.r 
file. This can be useful e.g. if you want to load a different file using 
\code{\link{gfd}}.
}
\details{
The name of the default metadata-file can be specified in the settings.
Only the experiment name can be changed via the \code{...} argument by providing 
the argument 'expName'. Other arguments will be ignored.
}
\examples{
\dontrun{
md <- getmd(); str(md); names(md)
md <- getmd("myFile.r")
md <- getmd(expName="OtherName") # to override the experiment name as specified 
# in the metadata.r file.
dataset <- gfd(getmd(expName="OtherName"))
}
}
\seealso{
\code{\link{metadata_file}}, \code{\link{getap}}, \code{\link{gdmm}}
}
