% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_noise.r
\name{noise_procedures}
\alias{noise_procedures}
\title{Record and apply noise}
\description{
Record a special noise-data file on your spectroscopy device, and 
use this file to calculate data that can then be used to specifically add 
'custom' noise to your datasets.
}
\details{
The noise data file is an R-data file generated / imported via the 
\code{\link{gfd}}) function, but containing only noise-spectra recorded 
following the recommendations below. This R-data file (where the factory-default 
filename is 'NoiseData') is then moved \strong{by the user} in to the AQUAP2SH 
folder. When generating datasets (a 'cube') using the function 
\code{\link{gdmm}}, it depends on whether you choose to add noise to the 
datasets or not (parameter \code{spl.do.noise} in the analysis procedure resp. 
your input. If, and only if, you choose to add noise, first (and only once per 
R-session) the noise R-data file is read in and used to calculate the specific 
noise distribution. This noise distribution is then used to specifically add 
noise to each individual dataset within the cube.
}
\seealso{
\code{\link{gdmm}}
}

