% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_noise.r
\name{noise_procedures}
\alias{noise_procedures}
\title{Record / add noise}
\description{
Record a special noise-data file on your spectroscopy device, and 
use this file to calculate data that can then be used to specifically add 
'custom' noise to your datasets. Alternatively, if you do not want to record 
a specific noise file, you can also add static noise (with mean=0) to your 
datasets. Use the parameter \code{noi_addMode} in the settings file to choose 
from one of the modes of adding noise, please see section 'Modes for noise 
calculation' for details.
}
\details{
The noise data file is an R-data file generated / imported via the 
\code{\link{gfd}}) function, but containing only noise-spectra recorded 
following the recommendations below. This R-data file (where the factory-default 
filename is 'NoiseData') is then moved \strong{by the user} in to the AQUAP2SH 
folder. When generating datasets (a 'cube') using the function 
\code{\link{gdmm}}, it depends on whether you choose to add noise to the 
datasets or not (parameter \code{spl.do.noise} in the analysis procedure resp. 
your input. If, and only if, you choose to add noise, first (and only once per 
R-session) the noise R-data file is read in and used to calculate the specific 
noise distribution. It is stored in an object where the name is starting with 
\code{nd_} (for 'noise distribution') in the environment \code{.ap2}; 
(\code{ls(.ap2)}). This noise distribution is then used to specifically add, 
according to the selected mode (see section 'Modes for noise calculation' 
below), noise to each single wavelength in each single observation in the 
individual dataset within the cube. If you do not want to use a specific 
noise-data file, you can set the noise mode to 'static' (parameter 
\code{noi_addMode} in the settings file).
}
\section{Procedure}{
 The procedure to work with noise-data files and use them to 
obtain a wavelength-specific noise distribution consists of the following 
steps:
\describe{
\item{Record noise spectra}{Use the function \code{\link{genNoiseRecExp}} to 
generate a folder structure for an experiment, then record the noise-spectra, 
for what the following procedure is suggested: 
XXX. 
Finally, use \code{\link{gfd}} to import the noise-data and create the R-data 
file (in the folder 'R-data' in the working directory of the experiment).}
\item{Move R-data file}{Move the resulting R-data file containing the noise-
spectra from the R-data folder into your \code{AQUAP2SH} folder, i.e. the 
folder also containing e.g. the settings.r file.}
\item{Choose to add noise to dataset}{In your actual experiment, in order to 
add noise to the dataset(s) you have to set the parameter \code{spl.do.noise} 
in the analysis procedure to \code{TRUE}. Alternatively, you can override the 
value from 'spl.do.noise' in the analysis procedure via the \code{...} 
argument in \code{\link{getap}} in \code{\link{gdmm}} - see examples.}
\item{Specify noise-data file}{In your actual experiment, specify the name of 
the file (residing in the folder \code{AQUAP2SH}) containing the noise-spectra 
either in the metadata or at the argument \code{noiseFile} in the function 
\code{\link{gdmm}} - see examples.}
\item{Evaluate difference}{Use any of the provided methods to check if there 
is a difference between the datasets without and with added noise.}
}
}

\section{Modes for noise calculation}{

Via the parameter \code{noi_addMode} in the settings file it can be specified 
which one of the available methods to calculate and add noise to the actual 
dataset should be used. The following values for \code{noi_addMode} are 
possible: 
\describe{
\item{sdNorm}{Factory default. Standard deviation based, 
sampling from \strong{normal} distribution. The noise distribution is described 
by \code{Noise_Mean +/- Noise_sd} whithin each single wavelength of the noise 
dataset. If noise should be added now to an actual dataset, for each single 
observation and there for each single wavelength the noise-value to be added 
to the actual spectral value can be described by 
\code{noiseValue = sample(rnorm(noi_sampleSize, mean=Noise_Mean, sd=Noise_sd),1)}. 
Additionally, all values outside the range of the noise-data in this specific 
wavelength get clipped. (\code{noi_sampleSize} is a parameter in the 
settings file.)}
\item{sdUnif}{Standard deviation based, 
sampling from \strong{uniform} distribution. The noise distribution is 
described by \code{Noise_Mean +/- Noise_sd} whithin each single wavelength of 
the noise dataset. If noise should be added now to an actual dataset, for each 
single observation and there for each single wavelength the noise-value to 
be added to the actual spectral value can be described by \code{noiseValue = 
sample(runif(noi_sampleSize, min=Noise_Mean-Noise_sd, max=Noise_Mean-Noise_sd),1)}. 
(\code{noi_sampleSize} is a parameter in the settings file.)}
\item{extrema}{The noise distribution simply represents the 
range of absorbance values within each single wavelength over all observations 
of the noise dataset. If noise should be added now to a dataset, for each single 
observation and there for each single wavelength a random process is selecting 
one of the two values from the corresponding wavelength in the noise 
distribution to be added to the original NIR data. In other words, for each 
single wavelength the original NIR-data absorbance values get increased or 
reduced by the maximum of (positive or negative) noise occuring at this 
specific wavelength.}
\item{static}{For this mode, no noise-data file is required. 
For every observation in the actual dataset and there for every single 
wavelength, a static noise normal-distribution is calculated using the 
parameters \code{noi_sampleSize} and \code{noi_staticValue} from the settings 
file: From the resulting pool of values then one is sampled and used to add to 
the spectral value of the actual dataset. 
(\code{noiseValue = sample(rnorm(noi_sampleSize, mean=0, 
sd=noi_staticValue),1)})}
}
}

\examples{
\dontrun{
fd <- gfd()
cu <- gdmm(fd, getap(spl.do.noise=TRUE))
cu <- gdmm(fd, getap(spl.do.noise=TRUE), noiseFile="FooBar") # instead of the 
# default noise-data filename resp. the filename specified in the metadata 
# file, use instead the noise-data file named 'FooBar'.
plot_pca(cu)
}
}
\seealso{
\code{\link{gdmm}}, \code{\link{settings_file}}, 
\code{\link{genFolderStr}}

Other Noise procedures: 
\code{\link{do_addNoise}()},
\code{\link{genNoiseRecExp}()}
}
\concept{Noise procedures}
