% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.r, R/plot_all.r
\docType{methods}
\name{plot,aquap_cube,ANY-method}
\alias{plot,aquap_cube,ANY-method}
\alias{plot_all_modells}
\alias{plot_cube}
\title{*** Plot All Available Models ***}
\usage{
\S4method{plot}{aquap_cube,ANY}(x, y, ...)

plot_cube(x, what = "all", ...)
}
\arguments{
\item{x}{An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.}

\item{y}{will be ignored}

\item{...}{Additional parameter to override the values of the graphical 
parameter in the analysis procedure file of the cube-object, see details.}

\item{what}{Character length one. What type of analysis to plot. Possible 
values are 'all', 'pca', 'sim', 'pls', 'aqg'.}
}
\value{
A PDF or graphic device.
}
\description{
Plot all available analysis graphics from the models in the 
cube-object. Function \code{plot} is a convenience function, it basically 
calls  \code{plot_cube}, what is the work-horse for plotting the 
cube-object.
}
\details{
The graphical parameters are taken from the analysis procedure 
stored in the cube-object. Via the \code{...} argument it is possible to 
override any of these parameters. Please see \code{\link{anproc_file}} for a 
complete listing and e.g. \code{\link{plot_pca_args}} and there the other 
functions of the 'Plot arguments' family for a separate listing of possible 
arguments.
}
\examples{
\dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot(cube, what="all") # the same as above
 plot(cube, what="pca")
}
}
\seealso{
\code{\link{plot_spectra}}

Other Core functions: \code{\link{exportSampleList}},
  \code{\link{gdmm}}, \code{\link{getFullData}}

Other Plot functions: \code{\link{plot,aquap_data,ANY-method}},
  \code{\link{plot_aqg}},
  \code{\link{plot_pca,aquap_cube-method}},
  \code{\link{plot_pls}}, \code{\link{plot_sim}}
}

