% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aqg.r
\name{plot_aqg}
\alias{plot_aqg}
\title{Plot Aquagram}
\usage{
plot_aqg(cube, ...)
}
\arguments{
\item{cube}{An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.}

\item{...}{Optional 'aqg' plotting parameters to override the values in the 
analysis procedure - for possible arguments see 
\code{\link{plot_aqg_args}}.}
}
\value{
A pdf or graphic device.
}
\description{
Plot Aquagram XXX.
}
\details{
The width and height of the resulting pdf can be set in the settings.
}
\examples{
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_aqg(cube)
 }
}
\seealso{
Other Aquagram documentation: \code{\link{calc_aqg_args}},
  \code{\link{plot_aqg_args}}

Other Plot functions: \code{\link{plot,aquap_cube,ANY-method}},
  \code{\link{plot,aquap_data,ANY-method}},
  \code{\link{plot_pca,aquap_cube-method}},
  \code{\link{plot_pls}}, \code{\link{plot_sim}}
}

