% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aqg.r
\name{plot_aqg}
\alias{plot_aqg}
\title{Plot Aquagram}
\usage{
plot_aqg(cube, aps = "def", ...)
}
\arguments{
\item{cube}{An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.}

\item{aps}{Character length one. The default way to obtain the analysis 
procedure containing the aquagram \strong{plotting} parameters. Defaults to 
"def". Possible values are:
\describe{
\item{"def"}{The default from the settings.r file is taken. (Argument 
\code{gen_plot_anprocSource})}
\item{"cube"}{Take the analysis procedure from within the cube, i.e. the 
analysis procedure that was used when creating the cube via \code{\link{gdmm}}
is used.}
\item{"defFile"}{Use the analysis procedure with the default filename as 
specified in the settings.r file in \code{fn_anProcDefFile}.}
\item{Custom filename}{Provide any valid filename for an analysis procedure to 
use as input for specifying the plotting options.}
}}

\item{...}{Optional 'aqg' plotting parameters to override the values in the 
analysis procedure - for possible arguments see 
\code{\link{plot_aqg_args}}.}
}
\value{
A pdf or graphic device.
}
\description{
Plot Aquagram XXX.
}
\details{
The width and height of the resulting pdf can be set in the settings.
}
\examples{
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_aqg(cube)
 plot_aqg(cube, aps="fooBar.r") # obtain the analysis procedure with the 
 # plotting parameters from the file 'fooBar.r'.
 }
}
\seealso{
Other Plot functions: \code{\link{plot,aquap_cube,missing-method}},
  \code{\link{plot,aquap_data,missing-method}},
  \code{\link{plot_da,aquap_cube-method}},
  \code{\link{plot_nnet,aquap_cube-method}},
  \code{\link{plot_pca,aquap_cube-method}},
  \code{\link{plot_pls,aquap_cube-method}},
  \code{\link{plot_pls_indepPred}},
  \code{\link{plot_rnf,aquap_cube-method}},
  \code{\link{plot_simca,aquap_cube-method}},
  \code{\link{plot_svm,aquap_cube-method}}

Other Aquagram documentation: \code{\link{calc_aqg_args}},
  \code{\link{plot_aqg_args}},
  \code{\link{tempCalib_procedures}}
}
