% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.r
\name{plot_pca}
\alias{plot_pca}
\title{@title Plot PCA
 @description Plot PCA scoresplots and / or loadings.
 @details The width and height of the resulting pdf can be set in the settings.
 @param cube An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.
 @param ... Optional pca plotting parameters to override the values in the 
 analysis procedure stored in the 'cube' - for possible arguments see 
 \code{\link{plot_pca_args}}.
 @return A pdf or graphic device.
 @family Plot functions
 @family PCA documentation
 @examples
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_pca(cube)
 }}
\usage{
plot_pca(cube, ...)
}
\description{
@title Plot PCA
 @description Plot PCA scoresplots and / or loadings.
 @details The width and height of the resulting pdf can be set in the settings.
 @param cube An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.
 @param ... Optional pca plotting parameters to override the values in the 
 analysis procedure stored in the 'cube' - for possible arguments see 
 \code{\link{plot_pca_args}}.
 @return A pdf or graphic device.
 @family Plot functions
 @family PCA documentation
 @examples
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_pca(cube)
 }
}

