% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.r, R/plot_pca.r
\docType{methods}
\name{plot_pca,aquap_cube-method}
\alias{plot_pca}
\alias{plot_pca,aquap_cube-method}
\alias{plot_pca,aquap_data-method}
\title{Plot PCA}
\usage{
\S4method{plot_pca}{aquap_cube}(object, ...)

\S4method{plot_pca}{aquap_data}(object, ...)
}
\arguments{
\item{object}{An object of class "aquap_cube" (as produced by 
\code{\link{gdmm}}) or of class "aqua_data" (as produced by 
\code{\link{gfd}}, see also \code{\link{getCubeDataset}}).}

\item{...}{Optional pca plotting parameters to override the values in the 
 analysis procedure - for possible arguments see 
 \code{\link{plot_pca_args}} and here below:
 \describe{
 \item{aps}{Character length one. The default way to obtain the analysis 
 procedure. Defaults to "def". Possible values are:
 \describe{
 \item{"def"}{The default from the settings.r file is taken. (Argument 
 \code{gen_plot_anprocSource})}
 \item{"cube"}{Take the analysis procedure from within the cube, i.e. the 
 analysis procedure that was used when creating the cube via \code{\link{gdmm}}
 is used.}
 \item{"defFile"}{Use the analysis procedure with the default filename as 
 specified in the settings.r file in \code{fn_anProcDefFile}.}
 \item{Custom filename}{Provide any valid filename for an analysis procedure to 
 use as input for specifying the plotting options.}
 }}
 \item{ld.bandwidth}{Character "def", or numeric length one. The bandwidth of 
wavelengths used in the peak-picking process when plotting the loadings. If 
left at the default "def", the value from the settings.r file 
(parameter \code{pp_bandwidth}) is used.}
 \item{ld.adLine}{Logical, numeric or character 'def'. If set to \code{FALSE},
no additional lines, if set to \code{TRUE} all the additional lines will be 
plotted in the loading plot. If an integer vector [2..5] is provided, one or 
more of the additional lines get plotted. See \code{\link{adLinesToVector}} 
for details. If left at the default value 'def', the value from the 
settings.r file (parameter \code{pca_AdLines}) is used.}
 \item{ld.col}{ Character "def" or a valid color vector. A possible custom color 
for coloring the loadings in the loading plot. When left at the default "def" 
the value fromt the settings.r file is read in (parameter 
\code{pca_ld_customColor}). Provide a color-vector to use these colors for 
coloring the loading plot.}
 \item{ld.lty}{Character "def" or a positive integer vector indicating the 
desired line types. When left at the default "def" the value from the 
settings.r file is read in (parameter \code{pca_ld_customLinetype}).}
 }}
}
\value{
A pdf or graphic device.
}
\description{
Plot PCA scoresplots and / or loadings.
}
\details{
By providing either an object of class "aquap_cube" or of class 
"aqua_data" to the argument \code{object}, you can plot either the pca-modells 
of the complete cube, or calculate and plot the pca-modell of a single dataset.
By providing one or more of the \code{...} arguments you can override the 
plotting values in the selected analysis procedure, see below.
The width and height of the resulting pdf can be set in the settings.
}
\examples{
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_pca(cube)
 plot_pca(cube, ld.bandwidth=12) # for detecting more peaks
 plot_pca(cube, ld.adLines=FALSE) # no additional lines in the loading plot 
 plot_pca(cube, ld.adLines=c(2, 4) # only vertical lines and WAMACs in the 
 # loading plot
plot_pca(cube, ld.col=c("red", "blue", "green"))
plot_pca(cube, ld.lty=c(1,2))
###
plot_pca(cube, aps="cube") # to take the analysis procedure from the cube
plot_pca(cube, aps="myAnproc.r") # take a custom analysis procedure file
###
fd <- gfd()
plot_pca(fd, pg.where="")
plot_pca(fd, pg.where="", pca.colorBy="C_Group") # assumes existence of "C_Group"
}
}
\seealso{
\code{\link{getCubeDataset}}

Other PCA documentation: \code{\link{calc_pca_args}},
  \code{\link{plot_pca_args}}

Other Plot functions: \code{\link{plot,aquap_cube,raw-method}},
  \code{\link{plot,aquap_data,ANY-method}},
  \code{\link{plot_aqg}}, \code{\link{plot_pls}},
  \code{\link{plot_sim}}
}

