% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plsr.r
\name{plot_pls}
\alias{plot_pls}
\title{Plot PLSR}
\usage{
plot_pls(cube, ...)
}
\arguments{
\item{cube}{An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.}

\item{...}{Optional 'pls' plotting parameters to override the values in the 
analysis procedure - for possible arguments see 
\code{\link{plot_pls_args}}.}
}
\value{
A pdf or graphic device.
}
\description{
Plot PLSR XXX.
}
\details{
The width and height of the resulting pdf can be set in the settings.
}
\examples{
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_pls(cube)
 }
}
\seealso{
Other PLSR documentation: \code{\link{calc_pls_args}},
  \code{\link{plot_pls_args}}

Other Plot functions: \code{\link{plot,aquap_cube,ANY-method}},
  \code{\link{plot,aquap_data,ANY-method}},
  \code{\link{plot_aqg}},
  \code{\link{plot_pca,aquap_cube-method}},
  \code{\link{plot_sim}}
}

