% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.r, R/plot_plsr.r
\docType{methods}
\name{plot_pls,aquap_cube-method}
\alias{plot_pls}
\alias{plot_pls,aquap_cube-method}
\title{Plot PLSR}
\usage{
\S4method{plot_pls}{aquap_cube}(object, ...)
}
\arguments{
\item{object}{An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.}

\item{...}{Optional 'pls' plotting parameters to override the values in the 
 analysis procedure - for possible arguments see \code{\link{plot_pls_args}} 
and below:
\describe{
\item{aps}{Character length one. The default way to obtain the analysis 
procedure. Defaults to "def". Possible values are:
\describe{
\item{"def"}{The default from the settings.r file is taken. (Argument 
\code{gen_plot_anprocSource})}
\item{"cube"}{Take the analysis procedure from within the cube, i.e. the 
analysis procedure that was used when creating the cube via \code{\link{gdmm}}
is used.}
\item{"defFile"}{Use the analysis procedure with the default filename as 
specified in the settings.r file in \code{fn_anProcDefFile}.}
\item{Custom filename}{Provide any valid filename for an analysis procedure to 
use as input for specifying the plotting options.}
}}
\item{rv.bandwidth}{Character "def", or numeric length one. The bandwidth of 
wavelengths used in the peak-picking process when plotting the regression 
vector. If left at the default "def", the value from the settings.r file 
(parameter \code{pp_bandwidth}) is used.}
 \item{rv.adLine}{Logical, numeric or character 'def'. If set to \code{FALSE},
no additional lines, if set to \code{TRUE} all the additional lines will be 
plotted in the loading plot. If an integer vector [2..5] is provided, one or 
more of the additional lines get plotted. See \code{\link{adLinesToVector}} 
for details. If left at the default value 'def', the value from the 
settings.r file (parameter \code{plsr_AdLines}) is used.}
 \item{rv.col}{ Character "def" or a valid color. A possible custom color 
for coloring the plsr regression vector. When left at the default "def" 
the value fromt the settings.r file is read in (parameter 
\code{plsr_rv_customColor}). Provide a color-vector to use it for 
coloring the plsr regression vector.}
 \item{rv.lty}{Character "def" or a positive integer indicating the 
desired line type. When left at the default "def" the value from the 
settings.r file is read in (parameter \code{plsr_rv_customLinetype}).}
}}
}
\value{
A pdf or graphic device.
}
\description{
Plot PLSR error and calibration / crossvalidation plots.
}
\details{
By providing one or more of the \code{...} arguments you can 
override the plotting values in the selected analysis procedure, see examples.
The width and height of the resulting pdf can be set in the settings.
}
\examples{
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_pls(cube)
 plot_pls(cube, aps="foobar.r") # to get the plotting parameters 
 # from the file "foobar.r".
 plot_pls(cube, lv.col="red") # to have the regression vector in red
 plot(cube, what="pls", pg.where="", rv.col="red", rv.lty=2)
 plot(cube, pg.where="", rv.col="red", rv.lty=2) 
 plot(cube, pls.what="regression") # for plsr, only plot the regression vector
}
}
\seealso{
Other PLSR documentation: \code{\link{calc_pls_args}},
  \code{\link{plot_pls_args}}

Other Plot functions: \code{\link{plot,aquap_cube,missing-method}},
  \code{\link{plot,aquap_data,missing-method}},
  \code{\link{plot_aqg}},
  \code{\link{plot_pca,aquap_cube-method}},
  \code{\link{plot_sim}}
}

