% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simca.r
\name{plot_sim}
\alias{plot_sim}
\title{Plot SIMCA}
\usage{
plot_sim(cube, ...)
}
\arguments{
\item{cube}{An object of class 'aquap_cube' as produced by \code{\link{gdmm}}.}

\item{...}{Optional 'sim' plotting parameters to override the values in the 
analysis procedure - for possible arguments see 
\code{\link{plot_sim_args}}.}
}
\value{
A pdf or graphic device.
}
\description{
Plot SIMCA interclass distances.
}
\details{
The width and height of the resulting pdf can be set in the settings.
}
\examples{
 \dontrun{
 dataset <- gfd()
 cube <- gdmm(dataset)
 plot(cube)
 plot_sim(cube)
}
}
\seealso{
Other Plot functions: \code{\link{plot,aquap_cube,missing-method}},
  \code{\link{plot,aquap_data,missing-method}},
  \code{\link{plot_aqg}},
  \code{\link{plot_pca,aquap_cube-method}},
  \code{\link{plot_pls}}

Other SIMCA documentation: \code{\link{calc_sim_args}},
  \code{\link{plot_sim_args}}
}

