% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.r, R/plot_spectra.r
\docType{methods}
\name{plot,aquap_data,missing-method}
\alias{plot,aquap_data,missing-method}
\alias{plot_spectra}
\alias{plot_spectra,aquap_cube-method}
\alias{plot_spectra,aquap_data-method}
\title{Plot Raw Spectra}
\usage{
\S4method{plot}{aquap_data,missing}(x, colorBy = NULL, ...)

\S4method{plot_spectra}{aquap_data}(x, colorBy = NULL, ...)

\S4method{plot_spectra}{aquap_cube}(x, colorBy = NULL, ...)
}
\arguments{
\item{x}{The standard dataset as produced by \code{\link{gfd}}, or a data-cube
as produced by \code{\link{gdmm}}.}

\item{colorBy}{Character vector, possible values are the class variables in 
the dataset. Provide a character vector with length > 1 to color along all 
these class variables; provide 'all' to use all available class variables for 
coloring - see examples.}

\item{...}{Optional general plotting options as defined in 
\code{\link{plot_pg_args}}.}
}
\description{
Plot the raw spectra contained in the dataset or in the cube.
}
\details{
If the provided object is of class \code{\link{aquap_cube}} and is 
containing more than one dataset, a single graphic with raw-spectra is 
produced for every dataset contained within the cube.
}
\examples{
\dontrun{
plot_spectra(dataset)
plot_spectra(dataset, colorBy="all")
plot(dataset) # the same as above
plot(dataset, colorBy=c("C_Group"))
plot_spectra(dataset, "C_Group")
plot_spectra(dataset, c("C_Group", "C_Repl")) # produces two plots
plot(dataset, pg.where="") # for plotting to graphic device
plot_spectra(dataset, c("C_Group", "C_Repl"), pg.where="", pg.main="foo")
cube <- gdmm(gfd())
plot_spectra(cube)
}
}
\seealso{
\code{\link{plot_all_modells}}

Other Core functions: \code{\link{exportSampleList}},
  \code{\link{gdmm}}, \code{\link{getFullData}},
  \code{\link{plot,aquap_cube,missing-method}}

Other Plot arguments: \code{\link{plot_aqg_args}},
  \code{\link{plot_pca_args}}, \code{\link{plot_pg_args}},
  \code{\link{plot_pls_args}}, \code{\link{plot_sim_args}}

Other Plot functions: \code{\link{plot,aquap_cube,missing-method}},
  \code{\link{plot_aqg}},
  \code{\link{plot_pca,aquap_cube-method}},
  \code{\link{plot_pls,aquap_cube-method}},
  \code{\link{plot_pls_indepPred}},
  \code{\link{plot_simca,aquap_cube-method}}
}

