% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.r, R/plot_classif_spec.r
\name{plot_svm,aquap_cube-method}
\alias{plot_svm,aquap_cube-method}
\alias{plot_svm}
\title{plot Discriminant Analysis}
\usage{
\S4method{plot_svm}{aquap_cube}(object, ...)
}
\arguments{
\item{object}{An object of class "aquap_cube" (as produced by 
\code{\link{gdmm}}).}

\item{...}{Optional \code{svm.} plotting parameters to override the values in 
the analysis procedure - for possible arguments see 
\code{\link{plot_SVM_args}} and here below: 
\describe{
\item{aps}{Character length one. The default way to obtain the analysis 
procedure. Defaults to "def". Possible values are:
\describe{
\item{"def"}{The default from the settings.r file is taken. (Argument 
\code{gen_plot_anprocSource})}
\item{"cube"}{Take the analysis procedure from within the cube, i.e. the 
analysis procedure that was used when creating the cube via \code{\link{gdmm}}
is used.}
\item{"defFile"}{Use the analysis procedure with the default filename as 
specified in the settings.r file in \code{fn_anProcDefFile}.}
\item{Custom filename}{Provide any valid filename for an analysis procedure to 
use as input for specifying the plotting options.}
}} 
}}
}
\description{
Plot all available models for discriminant analysis 
classification
}
\seealso{
Other Plot functions: 
\code{\link{plot,aquap_cube,missing-method}},
\code{\link{plot,aquap_data,missing-method}},
\code{\link{plot_aqg}()},
\code{\link{plot_classifX_indepPred}()},
\code{\link{plot_da,aquap_cube-method}},
\code{\link{plot_nnet,aquap_cube-method}},
\code{\link{plot_pca,aquap_cube-method}},
\code{\link{plot_pls,aquap_cube-method}},
\code{\link{plot_pls_indepPred}()},
\code{\link{plot_rnf,aquap_cube-method}},
\code{\link{plot_simca,aquap_cube-method}}

Other SVM documentation: 
\code{\link{calc_SVM_args}},
\code{\link{plot_SVM_args}}
}
\concept{Plot functions}
\concept{SVM documentation}
