% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{printStdColnames}
\alias{printStdColnames}
\title{Print standard column names}
\usage{
printStdColnames()
}
\description{
Prints the standard column names as defined in the local 
settings.r file to stdout.
}
\seealso{
Other Helper Functions: 
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{instAquap2Examples}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()},
\code{\link{updateAquap2}()},
\code{\link{updateSettings}()}
}
\concept{Helper Functions}
