% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_importData.r
\name{readHeader}
\alias{readHeader}
\title{Read Header}
\usage{
readHeader(md = getmd(), slType = "def", multiplyRows = "def")
}
\arguments{
\item{md}{List. The object with the metadat of the experiment. 
The default is to get the metadata file via  \code{\link{getmd}}.}

\item{slType}{Character. The type of sample-list file in the sampleLists/
sl_in folder. Possible values are:
\itemize{
   \item "def" Gets the default value from the settings.r file.
         (variable 'imp_sampleListType')
   \item NULL By providing 'NULL' to the argument 'slType' you indicate that 
         no sample-list should be imported to create the header. This would 
         be the case if you use a custom-function to import your spectral 
         data and all the necessary class- and numerical variables are already 
         defined in the same file that holds the spectral data. Please refer to
         \code{\link{custom_import}} for further information on the requirements 
         for this custom import function.
         You can use a custom function to import spectral data and at the same 
         time import additional variables from a sample-list file by providing 
         one of the characters listed below.
   \item "csv" a comma separated text file ending in '.csv'
   \item "txt" a tab delimited text file ending in '.txt'
   \item "xls" an Excel file ending in '.xlsx'
}}

\item{multiplyRows}{Character or Logical. If all the rows in the sample list 
should be multiplied by the number of consecutive scans as specified in the 
metadata of the experiment. 
If left at 'def' the value from the settings.r file is used. 
If 'FALSE' (what would be the case if you are using a sample list and, during 
your measurements, had to divert from the planned number of consecutive scans
or had to make other changes so that the number of consecutive scans is not 
the same for every sample) you have to manually insert the column (default name 
would be 'C_conSNr') into the sample list and provide the values for every row.
Set to 'TRUE' for multiplying every row in the sample list by the number of 
consecutive scans as specified in the metadata of the experiment.}
}
\description{
This functions reads in the sample-list file and prepares the 
header to be fused with the spectral data. 
Usually you do not need to call this function.
}
\details{
From the metadata, provided in the first argument, the experiment 
name is extracted; the sample list (what is used to create the header) must be in the 
sampleLists/sl_in folder and must be named with the experiment name, followed 
by a "-in" and then the file extension.
}
\examples{
\dontrun{
 header <- readHeader()
}
}
\seealso{
\code{\link{getFullData}}

Other Development Functions: \code{\link{custom_TRH}},
  \code{\link{custom_import}},
  \code{\link{imp_searchAskColumns}},
  \code{\link{readSpectra}}
}

