% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_importData.r
\name{readSpectra}
\alias{readSpectra}
\title{Read Spectra}
\usage{
readSpectra(md = getmd(), filetype = "def", naString = "NA", sh = NULL)
}
\arguments{
\item{md}{List. The object with the metadat of the experiment. 
The default is to get the metadata file via  \code{\link{getmd}}.}

\item{filetype}{Character. The type of the spectral raw data file. 
If a value other than "def" is provided, this is overriding the 
value of "filetype" in the metadata file. Possible values are:
\itemize{
\item "def": Gets the default value from the setings.r file. 
     (variable 'imp_specFileType')
\item "vision_NSAS.da": Import from the .da file generated by the Vision-software
from a Foss-XDS spectroscope.
\item "tabDelim.txt": Import any tab delimited text file that contains only 
the NIR spectra and *no* additional columns like e.g. time, temperature etc, 
and that has 1 character in front of the wavelengths in the column names of 
the NIR spectra.
\item "Pirouette.pir": Import spectra *and* any class- or numerical variable 
directly from a .pir file. Those column-names in the .pir  file that match the 
standard-column names (\code{\link{printStdColnames}}) as defined in the 
settings.r file will be assigned to those columns automatically.
\item "custom@yourFile.r": You can provide your own import-function for 
importing spectra, with "yourFile.r" being the .r-file located in the path 
specified in the .Renviron file. Please refer to \code{\link{custom_import}} 
for further information.
}}

\item{naString}{Character. What to use as 'NA'. Applies only when 'filetype' 
is 'tabDelim.txt'.}

\item{sh}{Character length one. Manual path to settings home. Can and should 
be left at the default \code{NULL}.}
}
\value{
The spectral data in the format as described in 'Value' in 
\code{\link{custom_import}}
}
\description{
Reads in just the spectra from the provided rawdata file.
}
\details{
This function is mainly for providing a possibility to test custom 
written rawdata import functions. (see \code{\link{custom_import}})
}
\seealso{
\code{\link{getFullData}}

Other Development Functions: 
\code{\link{custom_TRH}},
\code{\link{custom_import}},
\code{\link{imp_searchAskColumns}()},
\code{\link{readHeader}()}
}
\concept{Development Functions}
