% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prep_importData.r
\name{readSpectra}
\alias{readSpectra}
\title{Read Spectra}
\usage{
readSpectra(md = getmd(), filetype = "def", naString = "NA")
}
\arguments{
\item{md}{List. The object with the metadat of the experiment.
The default is to get the metadata file via  \code{\link{getmd}}.}

\item{filetype}{Character. The type of the spectral raw data file.
Possile values are:
\itemize{
\item "def": Gets the default value from the setings.r file.
     (variable 'imp_specFileType')
\item "vision_NSAS.da": Import from the .da file generated by the Vision-software
from a Foss-XDS spectroscope.
\item "tabDelim.txt": Import any tab delimited text file that contains only
the NIR spectra and *no* additional columns like e.g. time, temperature etc.
\item "custom@yourFile.r": You can provide your own import-function for
importing spectra, with "yourFile.r" being the .r-file located in the path
specified in the .Renviron file. Please refer to \code{\link{custom_import}}
for further information.
}}

\item{naString}{Character. What to use as 'NA'. Applies only when 'filetype'
is 'tabDelim.txt'.}
}
\value{
The spectral data in the format as described in 'Value' in
\code{\link{custom_import}}
}
\description{
Reads in just the spectra from the provided rawdata file.
}
\details{
This function is mainly for providing a possibility to test custom
written rawdata import functions. (see \code{\link{custom_import}})
}
\seealso{
\code{\link{getFullData}}

Other Development: \code{\link{custom_import}};
  \code{\link{readHeader}}
}

