% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prep_importData.r
\name{saveAQdata}
\alias{loadAQdata}
\alias{saveAQdata}
\title{Save and load aquap2 datasets}
\usage{
saveAQdata(dataset, md = getmd(), verbose = TRUE)

loadAQdata(md = getmd(), verbose = TRUE)
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{md}{List. The object with the metadat of the experiment.
The default is to get the metadata file via  \code{\link{getmd}}.}

\item{verbose}{Logical, if messages should be displayed.}
}
\value{
loadData
}
\description{
Save and load the standard aquap2 dataset (class "aquap_data")
to / from the 'R-data' folder.
}
\details{
From the provided metadata the experiment name is extracted.
\itemize{
 \item saveAQdata The dataset is saved under the same name as the experiment name.
 \item laodAQdata The file having the same name as the experiment name is being
 loaded.
}
}
\examples{
\dontrun{
saveAQdata(dataset)
loadAQdata()
}
}
\seealso{
\code{\link{getFullData}}

Other Helper Functions: \code{\link{genFolderStr}};
  \code{\link{updateSettings}}
}

