% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_generateDatasets.r
\name{selectWls}
\alias{selectWls}
\title{Select Wavelengths}
\usage{
selectWls(dataset, from, to)
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{from}{Numeric length one. The lower limit of the wavelengths to be 
selected.}

\item{to}{Numeric lengtho one. The upper limit of the wavelengths to be 
selected.}
}
\value{
The standard dataset (class 'aquap_data') containing only the 
selected wavelengths.
}
\description{
Select wavelengths from a dataset.
}
\details{
It is not necessary to exactly know the numbers of the wavelengths 
present in the dataset, as for the lower limit the same or the next highest 
wavelength as \code{from}, and for the upper limit the same or the next 
lowest wavelength as \code{to} will be selected.
}
\seealso{
\code{\link{siWl}}, \code{\link{siWlg}}

Other Data pre-treatment functions: \code{\link{[,aquap_data-method}},
  \code{\link{do_addNoise}}, \code{\link{do_avg}},
  \code{\link{do_detrend}}, \code{\link{do_emsc}},
  \code{\link{do_gapDer}}, \code{\link{do_msc}},
  \code{\link{do_resampleNIR}}, \code{\link{do_sgolay}},
  \code{\link{dpt_modules}}, \code{\link{ssc}}
}
