% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cube_generateDatasets.r
\name{selectWls}
\alias{selectWls}
\title{Select Wavelengths}
\usage{
selectWls(dataset, from, to)
}
\arguments{
\item{dataset}{An object of class 'aquap_data'}

\item{from}{Numeric length one. The lower limit of the wavelengths to be
selected.}

\item{to}{Numeric lengtho one. The upper limit of the wavelengths to be
selected.}
}
\value{
The standard dataset (class 'aquap_data')
}
\description{
Select wavelenghts from a dataset.
}
\details{
It is not necessary to exactly know the numbers of the wavelengths
present in the dataset, as for the lower limit the same or the next highest
wavelength as \code{from}, and for the upper limit the same or the next
lowest wavelength as \code{to} will be selected.
}
\seealso{
Other Data pre-treatment functions: \code{\link{do_msc}};
  \code{\link{do_sgolay}}; \code{\link{ssc}}
}

