% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cube_generateDatasets.r
\name{split_dataset}
\alias{split_dataset}
\title{Split Dataset}
\arguments{
\item{spl.var}{NULL or character vector. If NULL, no splitting of the
dataset will be performed. Provide a character vector with the column names
of class variables to split the dataset along these variables.}

\item{spl.wl}{NULL or character vector. If NULL, all in the dataset
available wavelengths will be used. Provide a character vector in the format
"wlFrom-to-wlTo" (e.g. c("1000-to-2000", "1300-to-1600", ...))
to use all previously defined splits in these wavelengths.}

\item{spl.do.smo}{Logical. If smoothing of the NIR data should be performed. The
values for the smoothing operation itself are defined in the settings.}

\item{spl.smo.raw}{Logical. If, should smoothing be performed, the raw
unsmoothed data will be used as well in addition to the smoothed data.}

\item{spl.do.noise}{Logical. If artifical noise should be added to the dataset.}

\item{spl.noise.raw}{If, should the noise-test be performed, the raw data
will be used as well in addition to the noise-data.}
}
\description{
The following parameters can be used in the \code{...} argument in
function \code{\link{getap}}, also within function \code{\link{gdmm}}, to
override the values in the analysis procedure file and so to modify the
split-dataset process - see examples.

\code{getap(...)}

\code{gdmm(dataset, ap=getap(...))}
}
\details{


For a list of all parameters that can be used in the \code{...}
argument in \code{\link{getap}} and in the \code{\link{plot}} functions
please see \code{\link{anproc_file}}.
}
\examples{
\dontrun{
dataset <- gfd() # will load or import data
cube <- gdmm(dataset, getap(spl.var="C_Group")) # split the dataset by "C_Group"
cube <- gdmm(dataset, getap(spl.var=c("C_Group", "C_Temp))) # split the dataset
# by "C_Group", then by "C_Temp"
cube <- gdmm(dataset, getap(spl.wl="1300-to-1600")) # override 'spl.wl' in the
# analysis procedure
}
}
\seealso{
Other Calc. arguments: \code{\link{calc_aqg_args}};
  \code{\link{calc_pca_args}}; \code{\link{calc_pls_args}};
  \code{\link{calc_sim_args}}
}

