% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_transformToAp2.r
\name{ttap}
\alias{ttap}
\alias{ttapM}
\title{Transform into aquap2 structure}
\usage{
ttap(obj, md = getmd(), dol = TRUE, ncpwl = 1, reduceTRH = FALSE, stf = TRUE)

ttapM(obj)
}
\arguments{
\item{obj}{The object to be converted}

\item{md}{The metadata; an object of class 'aquap_md'. expName can be
changed.}

\item{dol}{Logical. If outliers should be detected.}

\item{ncpwl}{The numbers of characters before the wavelength.}

\item{reduceTRH}{Logical. If Temp and RelHum values should be reduced.}

\item{stf}{Logical, if the resulting file should be saved in the r-data folder.}
}
\value{
An object of class 'aquap_data'
}
\description{
Transform a given object into the aquap2 data structure.
}
\details{
Class and Numeric prefixes in the column names are expected. 
Function \code{ttapM} is a reduced function for internal use only.
}
\concept{Integration Functions}
