% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen_general.r
\name{updateAquap}
\alias{updateAquap}
\title{Update the aquap2-package.}
\usage{
updateAquap(branch = "master")
}
\arguments{
\item{branch}{Character, the name of the branch to downlaod. Defaults to
"master".}
}
\description{
Download and install the latest version of package 'aquap2'
 from its github repository
}
\details{


Always downloads and installs the latest available version, also
 if the same up-to-date version is already installed.
}
\examples{
\dontrun{
 updateAquap()
 }
}
\seealso{
Other Helper Functions: \code{\link{genFolderStr}};
  \code{\link{loadAQdata}}, \code{\link{saveAQdata}};
  \code{\link{loadAquapDatapackage}};
  \code{\link{updateSettings}}
}

