% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniset_functions.R
\name{updateSettings}
\alias{updateSettings}
\title{Update Settings}
\usage{
updateSettings(silent = FALSE)
}
\arguments{
\item{silent}{Logical, defaults to FALSE.}
}
\description{
Update the Settings
}
\seealso{
Other Helper Functions: 
\code{\link{ap2_settings_setup}()},
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{getstn}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()}
}
\concept{Helper Functions}
