% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_general.r
\name{updateSettings}
\alias{updateSettings}
\title{Update aquap2 settings.}
\usage{
updateSettings(packageName = "aquap2", silent = FALSE)
}
\arguments{
\item{packageName}{Character, the name of the package where settings 
should be updated. Defaults to "aquap2".}

\item{silent}{Logical. If a confirmation should be printed. Defaults 
to 'FALSE'}
}
\value{
An (invisible) list with the settings resp. a list called 'stn' in 
the environment '.ap2'.
}
\description{
Manually read in the settings-file in the aquap2-settings 
home directory as specified in the .Renviron file.
}
\details{
If you leave 'autoUpdateSettings' in settings.r to 'TRUE', the 
settings will be checked resp. updated automatically every time you call any 
function from package 'aquap2'.
}
\section{Note}{
 If not present, the required `.Renviron` file will be 
automatically created. If the variable `AQUAP2SH` is not defined in the 
.Renviron file, it will be automatically added, and its default path is 
pointing to the (possibly also created) folder `aquap2SH` in the users home 
directory, where the `settings.r` file is automatically copied to if not 
already present. It is possible to manually provide a different path in the 
variable `AQUAP2SH` in the .Renviron file, pointing to any folder where then 
the settings.r file and all other relevant general files will reside.
}

\examples{
\dontrun{
updateSettings()
str(.ap2$stn)
ls(.ap2)
}
}
\seealso{
\code{\link{settings_file}}

Other Helper Functions: 
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{instAquap2Examples}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()},
\code{\link{updateAquap2}()}
}
\concept{Helper Functions}
