% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniset_functions.R
\name{updateSettings}
\alias{updateSettings}
\title{Update Settings}
\usage{
updateSettings(silent = FALSE)
}
\arguments{
\item{silent}{Logical, defaults to FALSE.}
}
\value{
An (invisible) list with the key=value pairs from the settings.R 
file.
}
\description{
Is first updating the settings.R file itself, and then is reading 
in the key=value pairs from the setting.R file.
}
\seealso{
Other Helper Functions: 
\code{\link{ap2_settings_setup}()},
\code{\link{ap2dme}()},
\code{\link{genFolderStr}()},
\code{\link{genNoiseRecExp}()},
\code{\link{genTempCalibExp}()},
\code{\link{getcd}()},
\code{\link{getcm}()},
\code{\link{getstn}()},
\code{\link{printStdColnames}()},
\code{\link{reColor}()},
\code{\link{saveAQdata}()},
\code{\link{siWlg}()}
}
\concept{Helper Functions}
