% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen_general.r
\name{updateSettings}
\alias{updateSettings}
\title{Update aquap2 settings.}
\usage{
updateSettings(packageName = "aquap2", silent = FALSE)
}
\arguments{
\item{packageName}{Character, the name of the package where settings
should be updated. Defaults to "aquap2".}

\item{silent}{Logical. If a confirmation should be printed. Defaults
to 'FALSE'}
}
\value{
An (invisible) list with the settings resp. a list called 'stn' in
the environment '.ap2'.
}
\description{
Manually read in the settings-file in the aquap2-settings
home directory as specified in the .Renviron file.
}
\details{
If you leave 'autoUpdateSettings' in settings.r to 'TRUE', the
settings will be checked resp. updated automatically every time you call any
function from package 'aquap2'.
}
\section{Note}{
 You have to set the path to where you want the settings.r file
to be stored once in your .Renviron file by defining
\code{AQUAP2SH = path/to/any/folder/XX} , with XX being any folder where then the
settings.r file will reside in. If you do not have a '.Renviron' file in your
home directory (user/home) you have to create one.
}
\examples{
\dontrun{
updateSettings()
str(.ap2$stn)
ls(.ap2)
}
}
\seealso{
\code{\link{settings_file}}

Other Helper Functions: \code{\link{genFolderStr}};
  \code{\link{loadAQdata}}, \code{\link{saveAQdata}};
  \code{\link{printStdColnames}}; \code{\link{updateAquap}}
}

